==> 01_Modeling FTs.Rmd <==
---
title: "Modeling FTs"
author: "Mallarie Yeager"
date: "2024-04-04"
output:
  pdf_document: default
  html_document: default
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
Load libraries and data
```{r}
library(readxl)
library(readr)
library(RColorBrewer)
library(visreg)
library(mgcv)
library(tidyr)
library(dplyr)

here::here()

#run functions
SEM <- function(x) sd(x)/sqrt(length(x))

#Set up color pal
wes_palettes <- list(
  Darjeeling1 = c("#FF0000", "#00A08A", "#F2AD00", "#F98400", "#5BBCD6"),
  Cavalcanti1 = c("#D8B70A", "#02401B", "#A2A475", "#81A88D", "#972D15"),
  GrandBudapest1 = c("#F1BB7B", "#FD6467", "#5B1A18", "#D67236"))
ponds <- c("NP","PP","PJ","WP", "GH", "QP")
pal <- data.frame(Pond = ponds, col = c(wes_palettes$Darjeeling1[1], 
                                        wes_palettes$Darjeeling1[2], 
                                        wes_palettes$Darjeeling1[5],
                                        wes_palettes$Darjeeling1[4], 
                                        wes_palettes$Cavalcanti1[1],
                                        wes_palettes$GrandBudapest1[3]))
##Load data
#functional traits file
morph_FTs <- read.csv("01_Modeling FTs/Data/All_FD_Data.csv")
Nutri_FTs <- read.csv("01_Modeling FTs/Data/FT_Nutrients_data.csv")
com.dat <- read.csv("01_Modeling FTs/Data/community_data_FT.csv")
```
Clean Dfs
```{r}
#Clean morphological FT df
morph_FTs <- morph_FTs[1:708,]
(num.spc <- table(morph_FTs$Common_name))
mean(num.spc)
SEM(num.spc)

#double check that all fish/size combos are unique - for merging nutrient data
#first subset fish that only have nutrient data for
morph_FTs.N <- subset(morph_FTs, morph_FTs$N.M == "N")
#look at fish/size combo and make sure numbers match from original Dat.1 data
fish <- unique(morph_FTs.N[c("Common_name", "FL..Image.J.")])
#morph_FTs.N = 200; fish = 200 check!
#remove unessecary columns for morphology FTs
morph_FTs <- morph_FTs[,-c(1:7,9,10,12)]
morph_FTs$FL..Image.J. <- as.numeric(morph_FTs$FL..Image.J.)
colnames(morph_FTs)[3] <- "FL"
morph_FTs <- morph_FTs[order(morph_FTs$Common_name),]

w <- which(is.na(rowSums(morph_FTs[ ,4:13]))== "TRUE")
morph_FTs <- morph_FTs[-w,]
morph_FTs <- morph_FTs[-c(145,280),]



#Bay Anchovy
for(i in 1:10){
plot(morph_FTs$FL[morph_FTs$Common_name == "Bay Anchovy"], morph_FTs[morph_FTs$Common_name == "Bay Anchovy", i + 3], main = colnames(morph_FTs)[i + 3])
}
tmp <- subset(morph_FTs, Common_name == "Bay Anchovy" )

morph_FTs$Common_name[morph_FTs$Common_name== "Cunner"] <- "Tautog" 
#com.dat.fish$Common_name[com.dat.fish$Common_name == "Cunner"] <- "Tautog" 
#pred.com.FT$Common_name[pred.com.FT$Common_name == "Cunner"] <- "Tautog" 

morph_FTs <- morph_FTs[-c(55,118,136,571),]
table(morph_FTs$Common_name)
morph_FTs <- morph_FTs[-which(morph_FTs$Common_name == "Atlantic Spot"),]
morph_FTs <- morph_FTs[-which(morph_FTs$Common_name == "Atlantic Mackerel"),]
morph_FTs <- morph_FTs[-which(morph_FTs$Common_name == "Bluefish"),]
morph_FTs <- morph_FTs[-which(morph_FTs$Common_name == "Threespine"),]
morph_FTs <- morph_FTs[-which(morph_FTs$Common_name == "Windowpane"),]
morph_FTs <- morph_FTs[-which(morph_FTs$Common_name == "Bluespotted Cornetfish"),]

uni.spc <- as.data.frame(unique(morph_FTs$Common_name))
colnames(uni.spc) <- "Common_name"
col <- colorRampPalette(brewer.pal(12,"Paired"))(nrow(uni.spc))
col <- cbind(as.character(uni.spc$Common_name), col)
colnames(col) <- c("Common_name", "col")
morph_FTs <- merge(morph_FTs, col, by = "Common_name")
morph_FTs$Common_name <- as.factor(morph_FTs$Common_name)

#Clean community survey df
#change cm to mm
com.dat$Length <- com.dat$Length * 10
colnames(com.dat)[7] <- "FL"
#remove inverts or fish not caught in FT sample
com.dat.fish <- merge(com.dat, uni.spc, by ="Common_name")

#Clean nutrient FT df
Nutri_FTs <- Nutri_FTs[, c(2,4,1,5,6,3,15,22)]
colnames(Nutri_FTs)[c(7:8)] <- c("Conc_Diff_NH4" ,"Conc_Diff_PO4")
Nutri_FTs <- na.omit(Nutri_FTs)
Nutri_FTs$NH4_PO4 <- Nutri_FTs$Conc_Diff_NH4 /Nutri_FTs$Conc_Diff_PO4
Nutri_FTs$NH4_PO4 <- ifelse(Nutri_FTs$NH4_PO4 == Inf, 0, Nutri_FTs$NH4_PO4)

colnames(Nutri_FTs)[5] <- "FL"
(uni.fam.col <- as.data.frame(unique(Nutri_FTs$Family)))
colnames(uni.fam.col) <- "Family"
col.fam <- colorRampPalette(brewer.pal(12,"Paired"))(nrow(uni.fam.col))
col.fam <- cbind(as.character(uni.fam.col$Family), col.fam)
colnames(col.fam) <- c("Family", "col.fam")
Nutri_FTs <- merge(Nutri_FTs, col.fam, by = "Family")
Nutri_FTs <- Nutri_FTs[order(Nutri_FTs$FL),]

table(Nutri_FTs$Family)
#remove fish where there is only one record
Nutri_FTs <- Nutri_FTs[-which(Nutri_FTs$Family == "Fistulariidae"),]
Nutri_FTs <- Nutri_FTs[-which(Nutri_FTs$Family == "Scombridae"),]
uni.fam <- as.data.frame(unique(c(Nutri_FTs[,1:2])))
colnames(uni.fam) <- c("Family", "Common_name")
tmp <- merge(com.dat.fish, uni.fam, by = "Common_name")
tmp1 <- as.data.frame(unique(tmp[,c(1,9)]))
com.dat.fish <- merge(com.dat.fish, tmp1, by = "Common_name")

#Empty FT df for predicting community FTs
pred.com.FT <- data.frame(com.dat.fish, B = NA, Osf = NA, Osh = NA, Ops = NA, Pro = NA, ES = NA, EP = NA, Bsh = NA, Bsf = NA, Pfp = NA, Cpt = NA, NH4 = NA, PO4 = NA, N_P = NA)
pred.com.FT <- merge(pred.com.FT, col, by = "Common_name")
```
Plot data to understand data dist shape- whether to transform 
```{r}
#check normaility of FTs
# par(mfrow = c(2,5))
# for(i in 1:10){
# tmp <- morph_FTs[ ,i + 3]
# hist(tmp, xlab = "", main = colnames(morph_FTs)[i+3])
# hist(log10(tmp), xlab = "", main = colnames(morph_FTs)[i+3])
# hist(log10(min(tmp[tmp > 0])/10 + tmp), xlab = "", main = colnames(morph_FTs)[i+3])
# print(shapiro.test(morph_FTs[ ,i + 3]))
# }
# sapply(morph_FTs[,3:13], FUN = min)
# 
# for(i in 1:10){
# hist(log10(1 + morph_FTs[ ,i + 3]), xlab = "", main = colnames(morph_FTs)[i+3])
#   tmp <- min(morph_FTs[ ,i + 3])
#   print(shapiro.test(log10(min(tmp[tmp > 0])/10 + morph_FTs[ ,i + 3])))
# }


min.FT <- apply(morph_FTs[,4:13], 2, FUN = function(x){min(x[x>0])})/10

# for(i in 1:10){
#   hist(min.FT[i] + morph_FTs[ ,i + 3], xlab = "", main = colnames(morph_FTs)[i+3])
# hist(log10(min.FT[i] + morph_FTs[ ,i + 3]), xlab = "", main = colnames(morph_FTs)[i+3])
# }

for(i in 1:10){ 
morph_FTs[,i + 3] <- log10(min.FT[i] + morph_FTs[,i + 3])
}
#for(i in 1:3){
#nutrients[,i + 9] <- log10(nutrients[,i + 9])
#}

# hist(Nutri_FTs$Conc_Diff_NH4)
# hist(Nutri_FTs$Conc_Diff_PO4)
# hist(Nutri_FTs$NH4_PO4)
# hist(log10(Nutri_FTs$Conc_Diff_NH4))
# hist(log10(Nutri_FTs$Conc_Diff_PO4))
# hist(log10(Nutri_FTs$NH4_PO4))

colnames(morph_FTs)[4:13] <- c("Osf", "Osh", "Ops", "Pro", "ES", "EP", "Bsh", "Bsf", "Pfp", "Cpt")

morph_FTs$Pro <- 10^(morph_FTs$Pro - min.FT[4])
morph_FTs$Biomass <- log10(morph_FTs$Biomass)

```

#Fitting GAMs for morphological traits
```{r}

#setting up newdata for predict funciton for morphology
newdata <- as.data.frame(matrix(ncol = nrow(uni.spc), nrow = 1000))
colnames(newdata) <- uni.spc$Common_name
for(i in 1:ncol(newdata)){
newdata[,i] <- seq(from = min(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]]), to = max(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]]), length.out =  1000)
}
newdata$rep <- 1:1000
newdata <- pivot_longer(newdata, cols = 1:length(newdata)-1,
               names_to = "Common_name",
               values_to = "FL")

newdata <- as.data.frame(newdata[,2:3])

####Biomass####
plot(morph_FTs$FL, morph_FTs$Biomass, col = as.character(morph_FTs$col), pch = 16)
hist(morph_FTs$Biomass)
mod.B <- gam(Biomass ~  Common_name + s(FL, by = Common_name, k = 18), method = "REML", data = morph_FTs)
summary(mod.B)
par(mfrow = c(2,2))
gam.check(mod.B)
par(mfrow = c(2,1))
visreg(mod.B, las = 2, xlab = "")



#Plot model over data
pred.Biomass <- predict(mod.B, newdata = newdata, interval = "conf")
pred.Biomass <- cbind(newdata[,1:2], pred.Biomass)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Biomass, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Biomass (log10(g))")
for(i in 1:nrow(uni.spc)){
lines(pred.Biomass$FL[pred.Biomass$Common_name == uni.spc[i,]], pred.Biomass$pred.Biomass[pred.Biomass$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

for(i in 1:15){ 
  
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Biomass[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"Biomass (log10)"))

  lines(pred.Biomass$FL[pred.Biomass$Common_name == uni.spc[i,]], pred.Biomass$pred.Biomass[pred.Biomass$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

####Osf####
mod.Osf <- gam(Osf ~ Common_name + s(FL, by = Common_name), method = "REML", data = morph_FTs)
summary(mod.Osf)
par(mfrow = c(2,2))
gam.check(mod.Osf)
par(mfrow = c(1,2))
visreg(mod.Osf)

#plot(mod.Osf, residuals = T)
#Plot model over data
pred.Osf <- predict(mod.Osf, newdata = newdata, interval = "conf")
pred.Osf <- cbind(newdata[,1:2], pred.Osf)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Osf, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Osf$FL[pred.Osf$Common_name == uni.spc[i,]], pred.Osf$pred.Osf[pred.Osf$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}
for(i in 1:nrow(uni.spc)){
  
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Osf[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"Osf "))

  lines(pred.Osf$FL[pred.Osf$Common_name == uni.spc[i,]], pred.Osf$pred.Osf[pred.Osf$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}


####Osh####
mod.Osh <- gam(Osh ~ Common_name + s(FL, by = Common_name), method = "REML", data = morph_FTs)
summary(mod.Osh)
k.check(mod.Osh)
par(mfrow = c(2,2))
gam.check(mod.Osh)
par(mfrow = c(1,2))
visreg(mod.Osh)
#Plot model over data
pred.Osh <- predict(mod.Osh, newdata = newdata, interval = "conf")
pred.Osh <- cbind(newdata[,1:2], pred.Osh)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Osh, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Osh$FL[pred.Osh$Common_name == uni.spc[i,]], pred.Osh$pred.Osh[pred.Osh$Common_name == uni.spc[i,]])
}
for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Osh[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"Osh (log10)"))

  lines(pred.Osh$FL[pred.Osh$Common_name == uni.spc[i,]], pred.Osh$pred.Osh[pred.Osh$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

####Ops####
mod.Ops <- gam(Ops ~ Common_name + s(FL, by = Common_name), method = "REML", data = morph_FTs)
summary(mod.Ops)
par(mfrow = c(2,2))
gam.check(mod.Ops)
par(mfrow = c(1,2))
visreg(mod.Ops)
#Plot model over data
pred.Ops <- predict(mod.Ops, newdata = newdata, interval = "conf")
pred.Ops <- cbind(newdata[,1:2], pred.Ops)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Ops , col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Ops$FL[pred.Ops$Common_name == uni.spc[i,]], pred.Ops$pred.Ops[pred.Ops$Common_name == uni.spc[i,]], 
      col = morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]])
}

for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Ops[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]],"Biomass (log10)")))

  lines(pred.Ops$FL[pred.Ops$Common_name == uni.spc[i,]], pred.Ops$pred.Ops[pred.Ops$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}


####Pro####
mod.Pro <- gam(Pro ~ Common_name + s(FL, by = Common_name), family = Tweedie(p = 1.1, link = "identity"), method = "REML", data = morph_FTs)
summary(mod.Pro)
par(mfrow = c(2,2))
gam.check(mod.Pro)
par(mfrow = c(1,2))
visreg(mod.Pro)
#Plot model over data
pred.Pro <- predict(mod.Pro, newdata = newdata, interval = "conf")
pred.Pro <- cbind(newdata[,1:2], pred.Pro)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Pro, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Pro$FL[pred.Pro$Common_name == uni.spc[i,]], pred.Pro$pred.Pro[pred.Pro$Common_name == uni.spc[i,]],
      col = morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]])
}

for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
      morph_FTs$Pro[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]],"Biomass (log10)")))

  lines(pred.Pro$FL[pred.Pro$Common_name == uni.spc[i,]], pred.Pro$pred.Pro[pred.Pro$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}


####ES####
mod.ES <- gam(ES ~ Common_name + s(FL, by = Common_name, k = 18), method = "REML",  data = morph_FTs)
summary(mod.ES)
par(mfrow = c(2,2))
gam.check(mod.ES)
par(mfrow = c(1,2))
visreg(mod.ES)
#Plot model over data
pred.ES <- predict(mod.ES, newdata = newdata, interval = "conf")
pred.ES <- cbind(newdata[,1:2], pred.ES)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$ES, col = as.character(morph_FTs$col), pch = 16)

for(i in 1:nrow(uni.spc)){
lines(pred.ES$FL[pred.ES$Common_name == uni.spc[i,]], pred.ES$pred.ES[pred.ES$Common_name == uni.spc[i,]])
}

for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$ES[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"ES (log10)"))

  lines(pred.ES$FL[pred.ES$Common_name == uni.spc[i,]], pred.ES$pred.ES[pred.ES$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}
####EP####
mod.EP <- gam(EP ~ Common_name + s(FL, by = Common_name, k = 18), method = "REML", data = morph_FTs)
summary(mod.EP)
par(mfrow = c(2,2))
gam.check(mod.EP)
par(mfrow = c(1,2))
visreg(mod.EP)
#Plot model over data
pred.EP <- predict(mod.EP, newdata = newdata, interval = "conf")
pred.EP <- cbind(newdata[,1:2], pred.EP)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$EP, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.EP$FL[pred.EP$Common_name == uni.spc[i,]], pred.EP$pred.EP[pred.EP$Common_name == uni.spc[i,]])
}

for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$EP[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"EP (log10)"))

  lines(pred.EP$FL[pred.EP$Common_name == uni.spc[i,]], pred.EP$pred.EP[pred.EP$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}
####Bsh####
mod.Bsh <- gam(Bsh ~ Common_name + s(FL, by = Common_name), method = "REML", data = morph_FTs)
summary(mod.Bsh)
par(mfrow = c(2,2))
gam.check(mod.Bsh)
par(mfrow = c(1,2))
visreg(mod.Bsh)
#Plot model over data
pred.Bsh <- predict(mod.Bsh, newdata = newdata, interval = "conf")
pred.Bsh <- cbind(newdata[,1:2], pred.Bsh)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Bsh, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Bsh$FL[pred.Bsh$Common_name == uni.spc[i,]], pred.Bsh$pred.Bsh[pred.Bsh$Common_name == uni.spc[i,]])
}

for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Bsh[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"Biomass (log10)"))

  lines(pred.Bsh$FL[pred.Bsh$Common_name == uni.spc[i,]], pred.Bsh$pred.Bsh[pred.Bsh$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

####Bsf####
mod.Bsf <- gam(Bsf ~ Common_name + s(FL, by = Common_name), method = "REML", data = morph_FTs)
summary(mod.Bsf)
par(mfrow = c(2,2))
gam.check(mod.Bsf)
par(mfrow = c(1,2))
visreg(mod.Bsf)
#Plot model over data
pred.Bsf <- predict(mod.Bsf, newdata = newdata, interval = "conf")
pred.Bsf <- cbind(newdata[,1:2], pred.Bsf)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Bsf, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Bsf$FL[pred.Bsf$Common_name == uni.spc[i,]], pred.Bsf$pred.Bsf[pred.Bsf$Common_name == uni.spc[i,]])
}


for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Bsf[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]],"Biomass (log10)")))

  lines(pred.Bsf$FL[pred.Bsf$Common_name == uni.spc[i,]], pred.Bsf$pred.Bsf[pred.Bsf$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}
####Pfp####
mod.Pfp <- gam(Pfp ~ Common_name + s(FL, by = Common_name), method = "REML", data = morph_FTs)
summary(mod.Pfp)
par(mfrow = c(2,2))
gam.check(mod.Pfp)
par(mfrow = c(1,2))
visreg(mod.Pfp)
#Plot model over data
pred.Pfp <- predict(mod.Pfp, newdata = newdata, interval = "conf")
pred.Pfp <- cbind(newdata[,1:2], pred.Pfp)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Pfp, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Pfp$FL[pred.Pfp$Common_name == uni.spc[i,]], pred.Pfp$pred.Pfp[pred.Pfp$Common_name == uni.spc[i,]])
}
for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Pfp[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]],"Biomass (log10)")))

  lines(pred.Pfp$FL[pred.Pfp$Common_name == uni.spc[i,]], pred.Pfp$pred.Pfp[pred.Pfp$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}
####Cpt####
mod.Cpt <- gam(Cpt ~ Common_name + s(FL, by = Common_name), method = "REML", data = morph_FTs)
summary(mod.Cpt)
par(mfrow = c(2,2))
gam.check(mod.Cpt)
par(mfrow = c(1,2))
visreg(mod.Cpt)
#Plot model over data
pred.Cpt <- predict(mod.Cpt, newdata = newdata, interval = "conf")
pred.Cpt <- cbind(newdata[,1:2], pred.Cpt)
par(mfrow = c(1,1))
plot(morph_FTs$FL, morph_FTs$Cpt, col = as.character(morph_FTs$col), pch = 16)
for(i in 1:nrow(uni.spc)){
lines(pred.Cpt$FL[pred.Cpt$Common_name == uni.spc[i,]], pred.Cpt$pred.Cpt[pred.Cpt$Common_name == uni.spc[i,]])
}

for(i in 1:nrow(uni.spc)){
  plot(morph_FTs$FL[morph_FTs$Common_name == uni.spc[i,]], 
       morph_FTs$Cpt[morph_FTs$Common_name == uni.spc[i,]], 
       col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]),
       pch = 16,
       xlab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]]),"FL"),
       ylab = paste(unique(morph_FTs$Common_name[morph_FTs$Common_name == uni.spc[i,]],"Biomass (log10)")))

  lines(pred.Cpt$FL[pred.Cpt$Common_name == uni.spc[i,]], pred.Cpt$pred.Cpt[pred.Cpt$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

```
#Fitting GAMs for morphological traits

```{r}
####Nutrients GAMS####
#setting up newdata for predict funciton of nutrients
tmp <- merge(Nutri_FTs, pred.com.FT, by = "Family")

uni.fam <- as.data.frame(unique(Nutri_FTs$Family))
uni.fam <- as.data.frame(uni.fam[order(uni.fam$`unique(Nutri_FTs$Family)`),])
colnames(uni.fam) <- "Family"
#colnames(newdata.n) <- uni.fam$Family

newdata.n <- as.data.frame(matrix(ncol = nrow(uni.fam), nrow = 1000))

for(i in 1:ncol(newdata.n)){
newdata.n[,i] <- seq(from = min(Nutri_FTs$FL[Nutri_FTs$Family == uni.fam[i,]]), to = max(Nutri_FTs$FL[Nutri_FTs$Family == uni.fam[i,]]), length.out = 1000)
}
colnames(newdata.n)<- uni.fam$Family
#newdata.n$rep <- 1:1000
newdata.n <- pivot_longer(newdata.n, cols = 1:length(newdata.n),
               names_to = "Family",
               values_to = "FL")

newdata.n <- as.data.frame(newdata.n)

Nutri_FTs$Family <- as.factor(Nutri_FTs$Family)
#NH4
mod.NH4 <- gam(Conc_Diff_NH4 ~ Family + s(FL, by = Family, k = 4), method = "REML", data = Nutri_FTs)
summary(mod.NH4)
par(mfrow = c(2,2))
#check_model(mod.NH4)
gam.check(mod.NH4)
par(mfrow = c(1,2))
visreg(mod.NH4)
#Plot model over data
pred.NH4 <- predict(mod.NH4, newdata = newdata.n, interval = "conf")
pred.NH4 <- cbind(newdata.n[,1:2], pred.NH4)
par(mfrow = c(1,1))
plot(Nutri_FTs$FL, Nutri_FTs$Conc_Diff_NH4, col = as.character(Nutri_FTs$col.fam), pch = 16)

for(i in 1:nrow(uni.fam)){
lines(pred.NH4$FL[pred.NH4$Family == uni.fam[i,1]], pred.NH4$pred.NH4[pred.NH4$Family == uni.fam[i,1]],
      col = as.character(Nutri_FTs$col.fam[Nutri_FTs$Family ==  uni.fam[i,1]]))
}

for(i in 1:nrow(uni.fam)){
  plot(Nutri_FTs$FL[Nutri_FTs$Family == uni.fam[i,]], 
       Nutri_FTs$Conc_Diff_NH4[Nutri_FTs$Family == uni.fam[i,]], 
       col = as.character(Nutri_FTs$col[Nutri_FTs$Family == uni.fam[i,]]),
       pch = 16,
       xlab = paste(unique(Nutri_FTs$Family[Nutri_FTs$Family == uni.fam[i,]]),"FL"),
       ylab = paste(unique(Nutri_FTs$Family[Nutri_FTs$Family == uni.fam[i,]]),"NH4"))

  lines(pred.NH4$FL[pred.NH4$Family == uni.fam[i,]], pred.NH4$pred.NH4[pred.NH4$Family == uni.fam[i,]],
      col = as.character(Nutri_FTs$col[Nutri_FTs$Family == uni.fam[i,]]))
}

boxplot(pred.NH4$pred.NH4 ~ pred.NH4$Family, las = 2)
abline(0,0)


#PO4
tmp <- aggregate(Conc_Diff_PO4 ~ Family, Nutri_FTs, min)

fams.zero <- subset(tmp, Conc_Diff_PO4 == 0, select = Family)
fams.no_zero <- subset(tmp, Conc_Diff_PO4 != 0, select = Family)

Nutri_FTs_no_zero <- merge(Nutri_FTs, fams.no_zero, by = "Family")
Nutri_FTs_zero <- merge(Nutri_FTs, fams.zero, by = "Family")

uni.fam <- as.data.frame(unique(Nutri_FTs_no_zero$Family))
uni.fam <- as.data.frame(uni.fam[order(uni.fam$`unique(Nutri_FTs_no_zero$Family)`),])
colnames(uni.fam) <- "Family"
newdata.p <- as.data.frame(matrix(ncol = nrow(uni.fam), nrow = 1000))

for(i in 1:ncol(newdata.p)){
newdata.p[,i] <- seq(from = min(Nutri_FTs_no_zero$FL[Nutri_FTs_no_zero$Family == uni.fam[i,]]), to = max(Nutri_FTs_no_zero$FL[Nutri_FTs_no_zero$Family == uni.fam[i,]]), length.out = 1000)
}
colnames(newdata.p)<- uni.fam$Family
#newdata.n$rep <- 1:1000
newdata.p <- pivot_longer(newdata.p, cols = 1:length(newdata.p),
               names_to = "Family",
               values_to = "FL")

newdata.p <- as.data.frame(newdata.p)

Nutri_FTs$Family <- as.factor(Nutri_FTs$Family)

fams.zero.PO4 <- Nutri_FTs_zero %>% 
  group_by(Family) %>%
  summarise(mean_PO4 = mean(Conc_Diff_PO4))

fams.mean.PO4 <- Nutri_FTs %>% 
  group_by(Family) %>%
  summarise(mean_PO4 = mean(Conc_Diff_PO4))
mod.PO4 <- gam(log10(Conc_Diff_PO4) ~ Family + s(FL, by=Family, k = 5), method = "REML",  data = Nutri_FTs_no_zero)

summary(mod.PO4)
par(mfrow = c(2,2))
 performance::check_model(mod.PO4)
gam.check(mod.PO4)
par(mfrow = c(2,1))
visreg(mod.PO4, las = 2, xlab = "")
#Plot model over data
pred.PO4 <- predict(mod.PO4, newdata = newdata.p, interval = "conf")
pred.PO4 <- cbind(newdata.p[,1:2], pred.PO4)
par(mfrow = c(1,1))
plot(Nutri_FTs_no_zero$FL, log10(Nutri_FTs_no_zero$Conc_Diff_PO4), col = as.character(Nutri_FTs_no_zero$col.fam), pch = 16)

for(i in 1:nrow(uni.fam)){
lines(pred.PO4$FL[pred.PO4$Family == uni.fam[i,1]], pred.PO4$pred.PO4[pred.PO4$Family == uni.fam[i,1]],
      col = as.character(Nutri_FTs_no_zero$col.fam[Nutri_FTs_no_zero$Family ==  uni.fam[i,1]]))
}


for(i in 1:nrow(uni.fam)){
  plot(Nutri_FTs_no_zero$FL[Nutri_FTs_no_zero$Family == uni.fam[i,]], 
       Nutri_FTs_no_zero$Conc_Diff_PO4[Nutri_FTs_no_zero$Family == uni.fam[i,]], 
       col = as.character(Nutri_FTs_no_zero$col[Nutri_FTs_no_zero$Family == uni.fam[i,]]),
       pch = 16,
       xlab = paste(unique(Nutri_FTs_no_zero$Family[Nutri_FTs_no_zero$Family == uni.fam[i,]]),"FL"),
       ylab = paste(unique(Nutri_FTs_no_zero$Family[Nutri_FTs_no_zero$Family == uni.fam[i,]]),"PO4"))

  lines(pred.PO4$FL[pred.PO4$Family == uni.fam[i,]], pred.PO4$pred.PO4[pred.PO4$Family == uni.fam[i,]],
      col = as.character(Nutri_FTs$col[Nutri_FTs_no_zero$Family == uni.fam[i,]]))
}

#N:P
Nutri_FTs$N_P_preGAM <- Nutri_FTs$NH4_PO4
mod.NH3_PO4 <- gam(N_P_preGAM ~ Family + s(FL, by=Family, k = 5), method = "REML", data = Nutri_FTs)

#mod.NH3_PO4 <- gam(N_P_preGAM ~ Family + s(FL, by=Family, k = 5),  family = Tweedie(p = 1.1, link = "identity"), method = "REML", data = Nutri_FTs)
summary(mod.NH3_PO4)
 par(mfrow = c(2,2))
 gam.check(mod.NH3_PO4)
 performance::check_model(mod.NH3_PO4)
 par(mfrow = c(1,2))
 visreg(mod.NH3_PO4)
# #Plot model over data
 pred.NH4_PO4 <- predict(mod.NH3_PO4, newdata = newdata.n, interval = "conf")
 pred.NH4_PO4 <- cbind(newdata.n, pred.NH4_PO4)
 par(mfrow = c(1,1))
plot(Nutri_FTs$FL, Nutri_FTs$N_P_preGAM, col = as.character(Nutri_FTs$col.fam), pch = 16)

 for(i in 1:nrow(uni.fam)){
lines(pred.NH4_PO4$FL[pred.NH4_PO4$Family == uni.fam[i,1]], pred.NH4_PO4$pred.NH4_PO4[pred.NH4_PO4$Family == uni.fam[i,1]],
      col = as.character(Nutri_FTs$col.fam[Nutri_FTs$Family ==  uni.fam[i,1]]))
}

```

Load model output into table
```{r}
gam.stats <- data.frame(mod =NA, family = NA, Rsq = NA, Dev_Expl = NA)
gam.stats[1,] <- cbind(paste(mod.B$formula[2], mod.B$formula[1], mod.B$formula[3]), mod.B$family[1], summary(mod.B)$r.sq, summary(mod.B)$dev.expl)
gam.stats[2,] <- cbind(paste(mod.Osf$formula[2],mod.Osf$formula[1], mod.Osf$formula[3]), mod.Osf$family[1], summary(mod.Osf)$r.sq, summary(mod.Osf)$dev.expl)
gam.stats[3,] <- cbind(paste(mod.Osh$formula[2],mod.Osh$formula[1], mod.Osh$formula[3]), mod.Osh$family[1], summary(mod.Osh)$r.sq, summary(mod.Osh)$dev.expl)
gam.stats[4,] <- cbind(paste(mod.Ops$formula[2],mod.Ops$formula[1], mod.Ops$formula[3]),  mod.Ops$family[1],  summary(mod.Ops)$r.sq, summary(mod.Ops)$dev.expl)
gam.stats[5,] <- cbind(paste(mod.Pro$formula[2],mod.Pro$formula[1], mod.Pro$formula[3]), mod.Pro$family[1],  summary(mod.Pro)$r.sq, summary(mod.Pro)$dev.expl)
gam.stats[6,] <- cbind(paste(mod.ES$formula[2],mod.ES$formula[1], mod.ES$formula[3]), mod.ES$family[1],   summary(mod.ES)$r.sq, summary(mod.ES)$dev.expl)
gam.stats[7,] <- cbind(paste(mod.EP$formula[2],mod.EP$formula[1], mod.EP$formula[3]), mod.EP$family[1], summary(mod.EP)$r.sq, summary(mod.EP)$dev.expl)
gam.stats[8,] <- cbind(paste(mod.Bsh$formula[2],mod.Bsh$formula[1], mod.Bsh$formula[3]), mod.Bsh$family[1],  summary(mod.Bsh)$r.sq, summary(mod.Bsh)$dev.expl)
gam.stats[9,] <- cbind(paste(mod.Bsf$formula[2],mod.Bsf$formula[1], mod.Bsf$formula[3]), mod.Bsf$family[1],  summary(mod.Bsf)$r.sq, summary(mod.Bsf)$dev.expl)
gam.stats[10,] <- cbind(paste(mod.Pfp$formula[2],mod.Pfp$formula[1], mod.Pfp$formula[3]), mod.Pfp$family[1],  summary(mod.Pfp)$r.sq, summary(mod.Pfp)$dev.expl)
gam.stats[11,] <-cbind(paste(mod.Cpt$formula[2],mod.Cpt$formula[1], mod.Cpt$formula[3]), mod.Cpt$family[1],   summary(mod.Cpt)$r.sq, summary(mod.Cpt)$dev.expl)
gam.stats[12,] <-cbind(paste(mod.NH4$formula[2],mod.NH4$formula[1], mod.NH4$formula[3]), mod.NH4$family[1],  summary(mod.NH4)$r.sq, summary(mod.NH4)$dev.expl)
gam.stats[13,] <-cbind(paste(mod.PO4$formula[2],mod.PO4$formula[1], mod.PO4$formula[3]), mod.PO4$family[1],  summary(mod.PO4)$r.sq, summary(mod.PO4)$dev.expl)
gam.stats[14,] <-cbind(paste(mod.NH3_PO4$formula[2],mod.NH3_PO4$formula[1], mod.NH3_PO4$formula[3]),  mod.NH3_PO4$family[1], summary(mod.NH3_PO4)$r.sq, summary(mod.NH3_PO4)$dev.expl)
gam.stats[,3:4] <- c(round(gam.stats$Rsq,2), round(gam.stats$Dev_Expl, 2))

```


```{r}
#Figure of model fit for Supplementary 
#Biomass GAM fit
png("/Users/mallarie.yeager/Documents/R_projects/FT Stability/Modeling FTs/Plots/gam_fit_plot.png",
    width = 800, height = 600, units = "px", bg = "white")
par(mfrow= c(3,4))
#plot(morph_FTs$FL, morph_FTs$Biomass, col = as.character(morph_FTs$col), pch = 16,
 #    xlab = "Fork Length (mm)", ylab = "Biomass (log10(g))")
#for(i in 1:nrow(uni.spc)){
#lines(pred.Biomass$FL[pred.Biomass$Common_name == uni.spc[i,]], pred.Biomass$pred.Biomass[pred.Biomass$Common_name == uni.spc[i,]],
#      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
#}

#Functional Traits GAM fits
#OSF
plot(morph_FTs$FL, morph_FTs$Osf, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Osf (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.Osf$FL[pred.Osf$Common_name == uni.spc[i,]], pred.Osf$pred.Osf[pred.Osf$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#OSH
plot(morph_FTs$FL, morph_FTs$Osh, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Osh (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.Osh$FL[pred.Osh$Common_name == uni.spc[i,]], pred.Osh$pred.Osh[pred.Osh$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#OPS
plot(morph_FTs$FL, morph_FTs$Ops, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Ops (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.Ops$FL[pred.Ops$Common_name == uni.spc[i,]], pred.Ops$pred.Ops[pred.Ops$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#PRO
plot(morph_FTs$FL, morph_FTs$Pro, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Pro")
for(i in 1:nrow(uni.spc)){
lines(pred.Pro$FL[pred.Pro$Common_name == uni.spc[i,]], pred.Pro$pred.Pro[pred.Pro$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#ES
plot(morph_FTs$FL, morph_FTs$ES, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "ES (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.ES$FL[pred.ES$Common_name == uni.spc[i,]], pred.ES$pred.ES[pred.ES$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#EP
plot(morph_FTs$FL, morph_FTs$EP, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "EP (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.EP$FL[pred.EP$Common_name == uni.spc[i,]], pred.EP$pred.EP[pred.EP$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#BSH
plot(morph_FTs$FL, morph_FTs$Bsh, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Bsh (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.Bsh$FL[pred.Bsh$Common_name == uni.spc[i,]], pred.Bsh$pred.Bsh[pred.Bsh$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#BSF
plot(morph_FTs$FL, morph_FTs$Bsf, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Bsf (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.Bsf$FL[pred.Bsf$Common_name == uni.spc[i,]], pred.Bsf$pred.Bsf[pred.Bsf$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#PFP
plot(morph_FTs$FL, morph_FTs$Pfp, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Pfp (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.Pfp$FL[pred.Pfp$Common_name == uni.spc[i,]], pred.Pfp$pred.Pfp[pred.Pfp$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}


#CPT
plot(morph_FTs$FL, morph_FTs$Cpt, col = as.character(morph_FTs$col), pch = 16,
     xlab = "Fork Length (mm)", ylab = "Cpt (log10)")
for(i in 1:nrow(uni.spc)){
lines(pred.Cpt$FL[pred.Cpt$Common_name == uni.spc[i,]], pred.Cpt$pred.Cpt[pred.Cpt$Common_name == uni.spc[i,]],
      col = as.character(morph_FTs$col[morph_FTs$Common_name == uni.spc[i,]]))
}

#NH4
plot(Nutri_FTs$FL, Nutri_FTs$Conc_Diff_NH4, col = as.character(Nutri_FTs$col.fam), pch = 16, xlab = "Fork Length (mm)", ylab = "NH4 (log10)")
for(i in 1:nrow(uni.fam)){
lines(pred.NH4$FL[pred.NH4$Family == uni.fam[i,1]], pred.NH4$pred.NH4[pred.NH4$Family == uni.fam[i,1]],
      col = as.character(Nutri_FTs$col.fam[Nutri_FTs$Family ==  uni.fam[i,1]]))
}

#PO4
plot(Nutri_FTs$FL, log10(Nutri_FTs$Conc_Diff_PO4), col = as.character(Nutri_FTs$col.fam), pch = 16, xlab = "Fork Length (mm)", ylab = "PO4 (log10)")
for(i in 1:nrow(uni.fam)){
lines(pred.PO4$FL[pred.PO4$Family == uni.fam[i,1]], pred.PO4$pred.PO4[pred.PO4$Family == uni.fam[i,1]],
      col = as.character(Nutri_FTs$col.fam[Nutri_FTs$Family ==  uni.fam[i,1]]))
}

#N:P
# plot(Nutri_FTs$FL, Nutri_FTs$N_P_preGAM, col = as.character(Nutri_FTs$col.fam), pch = 16, xlab = "Fork Length (mm)", ylab = "N:P (log10)")
# for(i in 1:nrow(uni.fam)){
# lines(pred.NH4_PO4$FL[pred.NH4_PO4$Family == uni.fam[i,1]], pred.NH4_PO4$pred.NH4_PO4[pred.NH4_PO4$Family == uni.fam[i,1]],
#       col = as.character(Nutri_FTs$col.fam[Nutri_FTs$Family ==  uni.fam[i,1]]))
# }
dev.off()
```


#Predicting Community Data
```{r}
pred.com.FT <- data.frame(com.dat.fish, B = NA, Osf = NA, Osh = NA, Ops = NA, Pro = NA, ES = NA, EP = NA, Bsh = NA, Bsf = NA, Pfp = NA, Cpt = NA, NH4 = NA, PO4 = NA, N_P_gam = NA)
pred.com.FT <- merge(pred.com.FT, col, by = "Common_name")

#subset sizes which only exist in morph_FTs

size.M <- dplyr::group_by(morph_FTs, Common_name) %>%
dplyr::summarise(min = min(FL), max = max(FL))
size.M[,2:3] <- round(size.M[,2:3])
size.M$min <- as.numeric(ifelse(nchar(size.M$min) < 3,
                                substr(size.M$min,1,1), 
                                substr(size.M$min,1,2)))*10 
size.M$max <- (as.numeric(ifelse(nchar(size.M$max) < 3, 
                                substr(size.M$max,1,1),
                                substr(size.M$max,1,2)))+1)*10

size.N <- dplyr::group_by(Nutri_FTs, Common_name) %>%
dplyr::summarise(min = min(FL), max = max(FL))
size.N[,2:3] <- round(size.N[,2:3])
size.N$min <- as.numeric(ifelse(nchar(size.N$min) < 3,
                                substr(size.N$min,1,1), 
                                substr(size.N$min,1,2)))*10 
size.N$max <- (as.numeric(ifelse(nchar(size.N$max) < 3, 
                                substr(size.N$max,1,1),
                                substr(size.N$max,1,2)))+1)*10

size <- rbind(size.M, size.N)

tmp <- aggregate(min ~ Common_name, size, max)
tmp2 <- aggregate(max ~ Common_name, size, min)
size <- merge(tmp, tmp2, by = "Common_name")
#colnames(size)[2:3] <- c("min", "max")
uni.spcs <- merge(size, uni.spc, by = "Common_name")

tmp3 <- NULL 
for(i in 1:nrow(uni.spcs)){
tmp <- subset(pred.com.FT, Common_name == uni.spcs[i,1])
tmp2 <- subset(tmp, FL < size$max[size$Common_name == uni.spcs[i,1]] & 
                 FL > size$min[size$Common_name == uni.spcs[i,1]])
tmp3 <- rbind(tmp3,tmp2)
}

pred.com.FT <- tmp3

par(mfrow = c(1,1))
pred.com.FT$B <- predict(mod.B, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$B, col = as.character(pred.com.FT$col), pch = 16, xlab = "Fork Length(mm)", ylab = "Biomass (log10(g))", main = "Predicted Biomass")
points(morph_FTs$FL, morph_FTs$Biomass, col = as.character(morph_FTs$col), pch = 16)

#backtransform biomass
pred.com.FT$B <- 10^pred.com.FT$B
plot(pred.com.FT$FL, pred.com.FT$B, col = as.character(pred.com.FT$col), pch = 16)

pred.com.FT$Osf <- predict(mod.Osf, newdata = pred.com.FT)

plot(pred.com.FT$FL, pred.com.FT$Osf, col = as.character(pred.com.FT$col), pch = 16 )
points(morph_FTs$FL, morph_FTs$Osf, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$Osh <- predict(mod.Osh, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$Osh, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$Osh, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$Ops <- predict(mod.Ops, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$Ops, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$Ops, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$Pro <- predict(mod.Pro, newdata = pred.com.FT)
pred.com.FT$Pro <- ifelse(pred.com.FT$Pro < 0, 0,pred.com.FT$Pro )
plot(pred.com.FT$FL, pred.com.FT$Pro, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$Pro, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$ES <- predict(mod.ES, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$ES, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$ES, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$EP <- predict(mod.EP, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$EP, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$EP, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$Bsh <- predict(mod.Bsh, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$Bsh, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$Bsh, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$Bsf <- predict(mod.Bsf, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$Bsf, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$Bsf, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$Pfp <- predict(mod.Pfp, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$Pfp, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$Pfp, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$Cpt <- predict(mod.Cpt, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$Cpt, col = as.character(pred.com.FT$col), pch = 16)
points(morph_FTs$FL, morph_FTs$Cpt, col = as.character(morph_FTs$col), pch = 1)

pred.com.FT$NH4 <- predict(mod.NH4, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$NH4,  col = as.character(pred.com.FT$col), pch = 16)
points(Nutri_FTs$FL, Nutri_FTs$Conc_Diff_NH4, col = Nutri_FTs$col.fam)


tmp <- merge(pred.com.FT, fams.zero, by = "Family")
tmp <- merge(tmp, fams.zero.PO4, by = "Family")
tmp$PO4 <- tmp$mean_PO4
tmp <- tmp[,-25]

tmp2 <- merge(pred.com.FT, fams.no_zero, by = "Family")
tmp2$PO4 <- predict(mod.PO4, newdata = tmp2)
plot(tmp2$FL, tmp2$PO4, col = as.character(tmp2$col), pch = 16, xlab = "Fork length (mm)", ylab = "PO4 (log10)", main = "Predicted Phosphate (PO4)")
points(Nutri_FTs_no_zero$FL, log10(Nutri_FTs_no_zero$Conc_Diff_PO4), col = Nutri_FTs$col.fam)
tmp2$PO4 <- 10^(tmp2$PO4)
pred.com.FT <- rbind(tmp, tmp2)
pred.com.FT <- pred.com.FT[order(pred.com.FT$Common_name),]

pred.com.FT$N_P_gam <- predict(mod.NH3_PO4, newdata = pred.com.FT)
plot(pred.com.FT$FL, pred.com.FT$N_P, col = as.character(pred.com.FT$col), pch = 16)
points(Nutri_FTs$FL, Nutri_FTs$N_P_PreGAM, col = Nutri_FTs$col.fam)

#remove fourspine outlier
#w <- which(pred.com.FT$N_P_gam > 17)
#pred.com.FT<- pred.com.FT[-w,]
#FT.cor <- cor(pred.com.FT.1[,11:23], use = "complete.obs")

boxplot(pred.com.FT[,c(11:23)], las = 2)
abline(0,0)
write.csv(pred.com.FT, "log_pred_FT_comm.csv")

pred.com.FT.1 <- pred.com.FT
for(i in 1:3){
pred.com.FT[,i + 10] <- 10^(pred.com.FT[,i + 10]- min.FT[i])
}
for(i in 5:10){
pred.com.FT[,i + 10] <- 10^(pred.com.FT[,i + 10]- min.FT[i])
}
#pred.com.FT.1[,21]  <- 10^(pred.com.FT[,21] - min.Nutri_FTs[1])
pred.com.FT[,23]  <- 10^(pred.com.FT[,23])

pred.com.FT$N_P <- pred.com.FT$NH4/pred.com.FT$PO4
plot(pred.com.FT$NH4, pred.com.FT$N_P, xlim =c(0,1100), ylim = c(0,1100), xlab = "Pre GAM N:P", ylab = "Post GAM N:P")
abline(0,1)
boxplot(pred.com.FT[,c(11:22,25)], las = 2)
abline(0,0)
write.csv(pred.com.FT[,c(1:22,25)], "/R Projects/FT Stability/Modeling FTs/Outputs/FT_community_df.csv", row.names = F)
write.csv(pred.com.FT[,c(1:22,25)], "/R Projects/FT Stability/Ordination Analyses/Data/FT_community_df.csv", row.names = F)

```


```{r}

##N_P gam
plot(pred.com.FT$N_P_gam, pred.com.FT$NH4)
plot(pred.com.FT$N_P_gam, pred.com.FT$PO4)



##N_P post gam
plot(pred.com.FT$N_P, pred.com.FT$NH4)
plot(pred.com.FT$N_P, pred.com.FT$PO4)
plot(pred.com.FT$NH4, pred.com.FT$PO4)

ggplot(pred.com.FT, aes(x = PO4, y = NH4, color = log10(N_P))) +
  geom_point(size = 2)+
  theme_bw()

```
