==> 03_Bray-curtis distance stability.Rmd <==
---
title: "Bray-Curtis distance Stability"
author: "Mallarie Yeager"
date: "2024-08-22"
output: html_document
editor_options: 
  chunk_output_type: console
---
```{r}
library(readxl)
library(readr)
library(RColorBrewer)
library(vegan)
library(FD)
library(ggplot2)
library(patchwork)
library(tidyverse)
library(ggplot2)
library(PNWColors)

here::here()

FT.sums <- read.csv("03_Bray Curtis distances_Stability/Data/Functional_trait_comm_dat.csv")
sums <- read.csv("03_Bray Curtis distances_Stability/Data/Species_comm_dat.csv")

#set color pal
wes_palettes <- list(
  Darjeeling1 = c("#FF0000", "#00A08A", "#F2AD00", "#F98400", "#5BBCD6"),
  Cavalcanti1 = c("#D8B70A", "#02401B", "#A2A475", "#81A88D", "#972D15"),
  GrandBudapest1 = c("#F1BB7B", "#FD6467", "#5B1A18", "#D67236")
)

ponds <- c("NP","PP","PJ","WP", "GH","QP")
pal <- data.frame(Pond = ponds,col = c(wes_palettes$Darjeeling1[1], 
                                        wes_palettes$Darjeeling1[2], 
                                        wes_palettes$Darjeeling1[5],
                                        wes_palettes$Darjeeling1[4], 
                                        wes_palettes$Cavalcanti1[1],
                                        wes_palettes$GrandBudapest1[3]))
```

```{r BETADISPERSER YR to Centroid}
#FT DATA
distmat.FT <- vegdist(FT.sums[,3:ncol(FT.sums)])
#calc yr-to-centroid
FT.beta <- betadisper(distmat.FT, group = FT.sums$Pond, type = "centroid")
FT.beta.dist <- as.data.frame(cbind(FT.sums[,1:2], FT.beta$distances))
FT.beta.dist$rel_distance <- FT.beta.dist$`FT.beta$distances`/max(FT.beta.dist$`FT.beta$distances`)
FT.beta.dist$Type <- "Functional Trait Composition"
colnames(FT.beta.dist)[3:4] <- c("distance", "rel_distance")

#calc yr-to-yr distance
dismat.FT.mat <- as.matrix(distmat.FT)
tmp.dist.FT <- c(dismat.FT.mat[1,2], dismat.FT.mat[2,3], dismat.FT.mat[3,4])
tmp.dist.FT <- NULL
for(i in 1:nrow(dismat.FT.mat) - 1){ 
tmp.dist.FT <- c(tmp.dist.FT, dismat.FT.mat[i,i+1])
}
tmp.dist.FT <- as.data.frame(tmp.dist.FT[-c(6,12,18,24,30)])
colnames(tmp.dist.FT) <- "distance"
tmp.dist.FT$rel_distance <- tmp.dist.FT$distance/max(tmp.dist.FT$distance)

#COMMUNITY DATA
distmat.com <- vegdist(sums[,3:ncol(sums)])
#calc yr-to-centroid

com.beta <- betadisper(distmat.com, group = sums$Pond, type = "centroid")
com.beta.dist <- as.data.frame(cbind(sums[,1:2], com.beta$distances))
com.beta.dist$rel_distance <- com.beta.dist$`com.beta$distances`/max(com.beta.dist$`com.beta$distances`)
com.beta.dist$Type <- "Community Composition"

colnames(com.beta.dist)[3:4] <- c("distance", "rel_distance")
#calc yr-to-yr distance
dismat.com.mat <- as.matrix(distmat.com)
tmp.dist.com <- NULL
for(i in 1:nrow(dismat.com.mat) - 1){ 
tmp.dist.com <- c(tmp.dist.com, dismat.com.mat[i,i+1])
}
tmp.dist.com <- as.data.frame(tmp.dist.com[-c(6,12,18,24,30)])
colnames(tmp.dist.com) <- "distance"
tmp.dist.com$rel_distance <- tmp.dist.com$distance/max(tmp.dist.com$distance)

#COMBINE FT and community data 
beta.yr_to_centroid_dist <-  rbind(com.beta.dist, FT.beta.dist)

ggplot(beta.yr_to_centroid_dist, aes(x = Year, y = distance, color = Pond))+
  geom_point()+
  facet_wrap(.~Type)

beta.yr_to_centroid_dist <- merge(beta.yr_to_centroid_dist, pal, by = "Pond")
summary(aov(rel_distance ~ Pond * Type, beta.yr_to_centroid_dist))
summary(aov(rel_distance ~ Pond, beta.yr_to_centroid_dist))
summary(aov(distance ~ Type, beta.yr_to_centroid_dist))

beta.yr_centroid_mean_pond <- aggregate(rel_distance ~ Pond, beta.yr_to_centroid_dist, mean)

beta.yr_centroid_mean_pond <- merge(beta.yr_centroid_mean_pond, pal, by = "Pond")

beta.yr_centroid_mean_type <- aggregate(rel_distance ~ Type, beta.yr_to_centroid_dist, mean)

beta.yr_to_yr_dist <- cbind(Pond = c(rep("GH",5), rep("NP",5), rep("PJ",5), rep("PP",5), rep("QP",5), rep("WP",5)), Year = rep(c("10_11", "11_12","12_13", "13_14", "14_15"),6),Type = c(rep("Community Composition",30), rep("Functional Trait Composition",30)), rbind(tmp.dist.com, tmp.dist.FT))

  
ggplot(beta.yr_to_yr_dist, aes(x = Year, y = distance, color = Pond))+
  geom_point()+
  facet_wrap(.~Type)

beta.yr_to_yr_dist <- merge(beta.yr_to_yr_dist, pal, by = "Pond")
summary(mod <- aov(rel_distance ~ Pond * Type, beta.yr_to_yr_dist))



summary(aov(rel_distance ~ Pond, beta.yr_to_yr_dist))
summary(aov(rel_distance ~ Type, beta.yr_to_yr_dist))
mod <- aov(rel_distance ~ Pond, beta.yr_to_yr_dist)
TukeyHSD(mod)

mod1 <- aov(rel_distance ~ Pond, beta.yr_to_centroid_dist)
TukeyHSD(mod1)
beta.yr_yr_mean_pond <- aggregate(rel_distance ~ Pond, beta.yr_to_yr_dist, mean)
beta.yr_yr_mean_pond <- merge(beta.yr_yr_mean_pond, pal, by = "Pond")

beta.yr_yr_mean_type <- aggregate(rel_distance ~ Type, beta.yr_to_yr_dist, mean)

par(mfrow = c(2,2))
##Main figs- Fig 1
#yr to centroid
##Pond
###Test
summary(aov.mod <- aov(rel_distance ~ Pond, beta.yr_to_centroid_dist))
TukeyHSD(aov.mod)
###Plot
boxplot(beta.yr_to_centroid_dist$rel_distance ~ Pond, beta.yr_to_centroid_dist, col = beta.yr_centroid_mean_pond$col, ylab = "Relative year to centroid distance")
points(beta.yr_centroid_mean_pond$rel_distance, pch = 16)
##Type
###Test
summary(aov.mod <- aov(rel_distance ~ Type, beta.yr_to_centroid_dist))
TukeyHSD(aov.mod)
###Plot
boxplot(beta.yr_to_centroid_dist$rel_distance ~ Type, beta.yr_to_centroid_dist,
        col = c("#506E82", "#768655"), ylab = "Relative year to centroid distance",
        names = c("Species", "Functional Traits"), xlab = "")
points(beta.yr_centroid_mean_type$rel_distance, pch = 16)
#yr to yr
##Pond
###Test
summary(aov.mod <- aov(rel_distance ~ Pond, beta.yr_to_yr_dist))
TukeyHSD(aov.mod)
###Plot
boxplot(beta.yr_to_yr_dist$rel_distance ~ Pond, beta.yr_to_yr_dist, col = beta.yr_yr_mean_pond$col, ylab = "Relative year to year distance")
points(beta.yr_yr_mean_pond$rel_distance, pch = 16)

###Test
summary(aov.mod <- aov(rel_distance ~ Type, beta.yr_to_yr_dist))
TukeyHSD(aov.mod)
yr.yr.ty.pval <- round(summary(aov.mod)[[1]][["Pr(>F)"]][1],3)
boxplot(beta.yr_to_yr_dist$rel_distance ~ Type, beta.yr_to_yr_dist,
        col = c("#506E82", "#768655"), ylab = "Relative year to year distance",
        names = c("Species", "Functional Traits"), xlab = "")
points(beta.yr_yr_mean_type$rel_distance, pch = 16)

```
set up for figure
```{r}
set.seed(12)
com.mds <- metaMDS(sums[,3:ncol(sums)],k=2, trymax = 10000, maxit = 10000)
coords.com.cent <- as.data.frame(com.mds$points)
coords.com.cent$Pond <- sums$Pond
coords.com.cent$Year <- sums$Year
coords.com.cent <- merge(coords.com.cent, pal, by = "Pond")


NP.e.C <- ordiellipse(com.mds, kind = "se", groups = sums$Pond, show = "NP", lwd = 2, col = as.character(coords.com.cent$col[coords.com.cent$Pond == "NP"]))
PP.e.C <- ordiellipse(com.mds, kind = "se",groups = sums$Pond, show = "PP", lwd = 2, col = as.character(coords.com.cent$col[coords.com.cent$Pond == "PP"]))
PJ.e.C <- ordiellipse(com.mds, kind = "se",groups = sums$Pond, show = "PJ", lwd = 2, col = as.character(coords.com.cent$col[coords.com.cent$Pond == "PJ"]))
WP.e.C <- ordiellipse(com.mds, kind = "se",groups = sums$Pond, show = "WP", lwd = 2, col= as.character(coords.com.cent$col[coords.com.cent$Pond == "WP"]))
GH.e.C <- ordiellipse(com.mds, kind = "se",groups = sums$Pond, show = "GH", lwd = 2, col = as.character(coords.com.cent$col[coords.com.cent$Pond == "GH"]))
QP.e.C <- ordiellipse(com.mds, kind = "se", groups = sums$Pond, show = "QP", lwd = 2, col = as.character(coords.com.cent$col[coords.com.cent$Pond == "QP"]))

coords.com.cent$MDS1centroid[coords.com.cent$Pond == "NP"] <- NP.e.C$NP$center[1]
coords.com.cent$MDS2centroid[coords.com.cent$Pond == "NP"] <- NP.e.C$NP$center[2]
coords.com.cent$MDS1centroid[coords.com.cent$Pond == "PP"] <- PP.e.C$PP$center[1]
coords.com.cent$MDS2centroid[coords.com.cent$Pond == "PP"] <- PP.e.C$PP$center[2]
coords.com.cent$MDS1centroid[coords.com.cent$Pond == "QP"] <- QP.e.C$QP$center[1]
coords.com.cent$MDS2centroid[coords.com.cent$Pond == "QP"] <- QP.e.C$QP$center[2]
coords.com.cent$MDS1centroid[coords.com.cent$Pond == "WP"] <- WP.e.C$WP$center[1]
coords.com.cent$MDS2centroid[coords.com.cent$Pond == "WP"] <- WP.e.C$WP$center[2]
coords.com.cent$MDS1centroid[coords.com.cent$Pond == "PJ"] <- PJ.e.C$PJ$center[1]
coords.com.cent$MDS2centroid[coords.com.cent$Pond == "PJ"] <- PJ.e.C$PJ$center[2]
coords.com.cent$MDS1centroid[coords.com.cent$Pond == "GH"] <- GH.e.C$GH$center[1]
coords.com.cent$MDS2centroid[coords.com.cent$Pond == "GH"] <- GH.e.C$GH$center[2]

FT.mds <- metaMDS(FT.sums[,3:ncol(FT.sums)], k = 2, trymax = 10000, maxit = 10000)
coords <- as.data.frame(FT.mds$points)
coords <- cbind(FT.sums[,1:2], coords)
coords <- merge(coords, pal, by = "Pond")

NP.e.F <- ordiellipse(FT.mds, kind = "se", groups = FT.sums$Pond, show = "NP", lwd = 2, col = as.character(coords$col[coords$Pond == "NP"]))
PP.e.F <- ordiellipse(FT.mds, kind = "se",groups = FT.sums$Pond, show = "PP", lwd = 2, col = as.character(coords$col[coords$Pond == "PP"]))
PJ.e.F <- ordiellipse(FT.mds, kind = "se",groups = FT.sums$Pond, show = "PJ", lwd = 2, col = as.character(coords$col[coords$Pond == "PJ"]))
WP.e.F <- ordiellipse(FT.mds, kind = "se",groups = FT.sums$Pond, show = "WP", lwd = 2, col= as.character(coords$col[coords$Pond == "WP"]))
GH.e.F <- ordiellipse(FT.mds, kind = "se",groups = FT.sums$Pond, show = "GH", lwd = 2, col = as.character(coords$col[coords$Pond == "GH"]))
QP.e.F <- ordiellipse(FT.mds, kind = "se", groups = FT.sums$Pond, show = "QP", lwd = 2, col = as.character(coords$col[coords$Pond == "QP"]))

coords$MDS1centroid[coords$Pond == "NP"] <- NP.e.F$NP$center[1]
coords$MDS2centroid[coords$Pond == "NP"] <- NP.e.F$NP$center[2]
coords$MDS1centroid[coords$Pond == "PP"] <- PP.e.F$PP$center[1]
coords$MDS2centroid[coords$Pond == "PP"] <- PP.e.F$PP$center[2]
coords$MDS1centroid[coords$Pond == "QP"] <- QP.e.F$QP$center[1]
coords$MDS2centroid[coords$Pond == "QP"] <- QP.e.F$QP$center[2]
coords$MDS1centroid[coords$Pond == "WP"] <- WP.e.F$WP$center[1]
coords$MDS2centroid[coords$Pond == "WP"] <- WP.e.F$WP$center[2]
coords$MDS1centroid[coords$Pond == "PJ"] <- PJ.e.F$PJ$center[1]
coords$MDS2centroid[coords$Pond == "PJ"] <- PJ.e.F$PJ$center[2]
coords$MDS1centroid[coords$Pond == "GH"] <- GH.e.F$GH$center[1]
coords$MDS2centroid[coords$Pond == "GH"] <- GH.e.F$GH$center[2]

```


###Figure 1 multipanel Plot
```{r}


jpeg("03_Bray Curtis distances_Stability/Plots/Community_vs_Functiontial_trait_stability_plot.jpeg",width = 12, height = 24, units = "cm", res = 600, pointsize = 10)
par(mfrow =c(4,2))
#1
par(mar=c(4, 5, 1.5, 0.5))
plot(coords.com.cent$MDS1, coords.com.cent$MDS2,  type = "n", xlab = "nMDS1", xlim =  c(-1.1,0.7), 
     ylim = c(-0.8,0.7), ylab = "Species Composition \nnMDS2", main = "Year to Centroid")
abline(h = 0, lty = 2)
abline(v = 0, lty = 2)
for (i in 1:nrow(coords.com.cent)) {
  text(coords.com.cent$MDS1[i], coords.com.cent$MDS2[i], 
       labels = substr(as.character(coords.com.cent$Year[i]), 
                       start = 3, stop = 4),
       col = c(as.character(coords.com.cent$col[i])),
       pos = 1, cex = 0.75)
}
points(NP.e.C$NP$center[1], NP.e.C$NP$center[2], col = as.character(pal$col[pal$Pond == "NP"]), pch = 16)
points(PP.e.C$PP$center[1], PP.e.C$PP$center[2], col = as.character(pal$col[pal$Pond == "PP"]), pch = 16)
points(QP.e.C$QP$center[1], QP.e.C$QP$center[2], col = as.character(pal$col[pal$Pond == "QP"]), pch = 16)
points(PJ.e.C$PJ$center[1], PJ.e.C$PJ$center[2], col = as.character(pal$col[pal$Pond == "PJ"]), pch = 16)
points(GH.e.C$GH$center[1], GH.e.C$GH$center[2], col = as.character(pal$col[pal$Pond == "GH"]), pch = 16)
points(WP.e.C$WP$center[1], WP.e.C$WP$center[2], col = as.character(pal$col[pal$Pond == "WP"]), pch = 16)
for (j in 1:length(ponds)) {
  
  s <- subset(coords.com.cent, Pond == ponds[j])
  s <- s[order(s$Year), ]
  #points(s$MDS1, s$MDS2, pch = 21, col = as.character(s$col[1]), bg = s$col[1], cex = 0.5)
  for (k in 1:(nrow(s))) {
    arrows(x0 = s$MDS1[k], y0 = s$MDS2[k],
           x1 = s$MDS1centroid[k], y1 =s$MDS2centroid[k] ,
           code = 1, col = as.character(s$col[k]), len = 0)
  }
}
legend("topright", legend = c("NP", "PP", "PJ", "WP", "GH", "QP"),
       col = as.character(pal$col), ncol = 2, lwd = 2, lty = 1, bty = "n", cex = 0.5)

legend("bottomright", legend = paste( "Stress = ", round(com.mds$stress,2)),
       bty = "n", text.col = "grey")
legend(-1.4,0.8, legend = "(a)", bty = "n", cex = 1.4)


#2
par(mar=c(4, 4.5, 1.5, 0.5))
plot(coords.com.cent$MDS1, coords.com.cent$MDS2,  type = "n", xlab = "nMDS1", xlim =  c(-1.1,0.7), ylim = c(-0.8,0.7), ylab = "nMDS2", main  = "Year to Year")
abline(h = 0, lty = 2)
abline(v = 0, lty = 2)
for (i in 1:nrow(coords.com.cent)) {
  text(coords.com.cent$MDS1[i], coords.com.cent$MDS2[i], 
       labels = substr(as.character(coords.com.cent$Year[i]), 
                       start = 3, stop = 4),
       col = c(as.character(coords.com.cent$col[i])),
       pos = 1, cex = 0.75)
}
for (j in 1:length(ponds)) {
  s <- subset(coords.com.cent, Pond == ponds[j])
  s <- s[order(s$Year), ]
  points(s$MDS1, s$MDS2, pch = 21, col = as.character(s$col[1]), bg = s$col[1], cex = 0.5)
  for (k in 1:(nrow(s) - 1)) {
    arrows(x0 = s$MDS1[k], y0 = s$MDS2[k],
           x1 = s$MDS1[k+1], y1 = s$MDS2[k+1],
           code = 2, col = as.character(s$col[k]), len = 0.1)
  }
}
legend("topright", legend = c("NP", "PP", "PJ", "WP", "GH", "QP"),
       col = as.character(pal$col), ncol = 2, lwd = 2, lty = 1, bty = "n", cex = 0.5)

legend("bottomright", legend = paste( "Stress = ", round(com.mds$stress,2)),
       bty = "n", text.col = "grey")
legend(-1.4,0.8, legend = "(b)", bty = "n", cex = 1.4)

#3
par(mar=c(4, 5, 1.5, 0.5))
plot(coords$MDS1, coords$MDS2,  type = "n", xlab = "nMDS1" ,ylab = "Fuctional Trait Composition\n nMDS2",
     xlim = c(-0.4,0.4), ylim = c(-0.2,0.2))
abline(h = 0, lty = 2)
abline(v = 0, lty = 2)
for (i in 1:nrow(coords)) {
  text(coords$MDS1[i], coords$MDS2[i], 
       labels = substr(as.character(coords$Year[i]), 
                       start = 3, stop = 4),
       col = c(as.character(coords$col[i])),
       pos = 1, cex = 0.75)
}
points(NP.e.F$NP$center[1], NP.e.F$NP$center[2], col = as.character(pal$col[pal$Pond == "NP"]), pch = 16)
points(PP.e.F$PP$center[1], PP.e.F$PP$center[2], col = as.character(pal$col[pal$Pond == "PP"]), pch = 16)
points(QP.e.F$QP$center[1], QP.e.F$QP$center[2], col = as.character(pal$col[pal$Pond == "QP"]), pch = 16)
points(PJ.e.F$PJ$center[1], PJ.e.F$PJ$center[2], col = as.character(pal$col[pal$Pond == "PJ"]), pch = 16)
points(GH.e.F$GH$center[1], GH.e.F$GH$center[2], col = as.character(pal$col[pal$Pond == "GH"]), pch = 16)
points(WP.e.F$WP$center[1], WP.e.F$WP$center[2], col = as.character(pal$col[pal$Pond == "WP"]), pch = 16)


for (j in 1:length(ponds)) {
  s <- subset(coords, Pond == ponds[j])
  s <- s[order(s$Year), ]
  #points(s$MDS1, s$MDS2, pch = 21, col = as.character(s$col[1]), bg = s$col[1], cex = 0.5)
  for (k in 1:(nrow(s))) {
    arrows(x0 = s$MDS1[k], y0 = s$MDS2[k],
           x1 = s$MDS1centroid [k], y1 =s$MDS2centroid[k] ,
           code = 1, col = as.character(s$col[k]), len = 0)
  }
}
legend("topright", legend = pal$Pond,
       col = as.character(pal$col), ncol = 2, lwd = 2, lty = 1, bty = "n", cex = 0.5)
legend("bottomright", legend = paste( "Stress = ", round(FT.mds$stress,2)),
       bty = "n", text.col = "grey")
legend(-0.53,0.23, legend = "(c)", bty = "n", cex = 1.4)


#4
par(mar=c(4, 4.5, 1.5, 0.5))
plot(coords$MDS1, coords$MDS2,  type = "n", xlab = "nMDS1" ,ylab = "nMDS2",
     xlim = c(-0.4,0.4), ylim = c(-0.2,0.2))
abline(h = 0, lty = 2)
abline(v = 0, lty = 2)
for (i in 1:nrow(coords)) {
  text(coords$MDS1[i], coords$MDS2[i], 
       labels = substr(as.character(coords$Year[i]), 
                       start = 3, stop = 4),
       col = c(as.character(coords$col[i])),
       pos = 1, cex = 0.75)
}
for (j in 1:length(ponds)) {
  s <- subset(coords, Pond == ponds[j])
  s <- s[order(s$Year), ]
  points(s$MDS1, s$MDS2, pch = 21, col = as.character(s$col[1]), bg = s$col[1], cex = 0.5)
  for (k in 1:(nrow(s) - 1)) {
    arrows(x0 = s$MDS1[k], y0 = s$MDS2[k],
           x1 = s$MDS1[k+1], y1 = s$MDS2[k+1],
           code = 2, col = as.character(s$col[k]), len = 0.1)
  }
}

legend("topright", legend = pal$Pond,
       col = as.character(pal$col), ncol = 2, lwd = 2, lty = 1, bty = "n", cex = 0.5)
legend("bottomright", legend = paste( "Stress = ", round(FT.mds$stress,2)),
       bty = "n", text.col = "grey")
legend(-0.53,0.23, legend = "(d)", bty = "n", cex = 1.4)

#5
summary(mod <- aov(rel_distance ~ Pond, beta.yr_to_centroid_dist ))
boxplot(beta.yr_to_centroid_dist$rel_distance ~ Pond, beta.yr_to_centroid_dist, col = beta.yr_centroid_mean_pond$col, ylab = "Relative year to centroid distance")
points(beta.yr_centroid_mean_pond$rel_distance, pch = 16)
legend(-0.45,1.07, legend = "(e)", bty = "n", cex = 1.4)
legend("topright", legend = "P = 0.006", bty = "n", cex = 0.8)

#6
summary(mod <- aov(rel_distance ~ Pond,beta.yr_to_yr_dist ))
boxplot(beta.yr_to_yr_dist$rel_distance ~ Pond, beta.yr_to_yr_dist, col = beta.yr_yr_mean_pond$col, ylab = "Relative year to year distance")
points(beta.yr_yr_mean_pond$rel_distance, pch = 16)
legend(-0.45,1.07, legend = "(f)", bty = "n", cex = 1.4)
legend("topright", legend = "P = 0.07", bty = "n", cex = 0.8)

#7
summary(mod <- aov(rel_distance ~ Type, beta.yr_to_centroid_dist ))
boxplot(beta.yr_to_centroid_dist$rel_distance ~ Type, beta.yr_to_centroid_dist,
        col = c("#506E82", "#768655"), ylab = "Relative year to centroid distance",
        names = c("Species", "Functional Traits"), xlab = "Composition Type")
points(beta.yr_centroid_mean_type$rel_distance, pch = 16)
legend(0.2,1.07, legend = "(g)", bty = "n", cex = 1.4)
legend("topright", legend = "P = 0.03", bty = "n", cex = 0.8)

#8
summary(mod <- aov(rel_distance ~ Type,beta.yr_to_yr_dist ))
boxplot(beta.yr_to_yr_dist$rel_distance ~ Type, beta.yr_to_yr_dist,
        col = c("#506E82", "#768655"), ylab = "Relative year to year distance",
        names = c("Species", "Functional Traits"), xlab = "Composition Type")
points(beta.yr_yr_mean_type$rel_distance, pch = 16)
legend(0.2,1.07, legend = "(h)", bty = "n", cex = 1.4)
legend("topright", legend = "P = 0.003", bty = "n", cex = 0.8)
dev.off()
```
