==> 04_Simper analysis and biplot.Rmd <==
---
title: "SIMPER analysis and Biplot"
author: "Mallarie Yeager"
date: "2024-08-22"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
Load libraries and data
```{r}
library(readxl)
library(readr)
library(RColorBrewer)
library(vegan)
library(FD)
library(ggplot2)
library(patchwork)
library(tidyverse)
library(ggplot2)
library(PNWColors)

here::here()
#load data
FT.sums <- read.csv("04_ANCOVA_SIMPER_BIPOLT Analyses/Data/Functional_trait_comm_dat.csv")
sums <- read.csv("04_ANCOVA_SIMPER_BIPOLT Analyses/Data/Species_comm_dat.csv")
TL <- read.csv("04_ANCOVA_SIMPER_BIPOLT Analyses/Data/fish_diets.csv")
com.dat.fish <- read.csv("02_Ordination Analysis/Data/com.dat.fish.csv")
#set color pal
wes_palettes <- list(
  Darjeeling1 = c("#FF0000", "#00A08A", "#F2AD00", "#F98400", "#5BBCD6"),
  Cavalcanti1 = c("#D8B70A", "#02401B", "#A2A475", "#81A88D", "#972D15"),
  GrandBudapest1 = c("#F1BB7B", "#FD6467", "#5B1A18", "#D67236")
)

ponds <- c("NP","PP","PJ","WP", "GH","QP")
pal <- data.frame(Pond = ponds,col = c(wes_palettes$Darjeeling1[1], 
                                        wes_palettes$Darjeeling1[2], 
                                        wes_palettes$Darjeeling1[5],
                                        wes_palettes$Darjeeling1[4], 
                                        wes_palettes$Cavalcanti1[1],
                                        wes_palettes$GrandBudapest1[3]))
colnames(sums)[3:ncol(sums)] <- gsub(".", " ", colnames(sums[,3:ncol(sums)]), fixed = T)
```
PERMANOVA
```{r}
#PERMANOVA
adonis2(FT.sums[,3:ncol(FT.sums)] ~ Pond * as.numeric(Year), FT.sums, distance = "Bray")
adonis2(FT.sums[,3:ncol(FT.sums)] ~ Pond, FT.sums, distance = "Bray")
adonis2(FT.sums[,3:ncol(FT.sums)] ~ as.numeric(Year), FT.sums, distance = "Bray")
adonis2(sums[,3:ncol(sums)] ~ Pond *  as.numeric(Year), sums, distance = "Bray")
adonis2(sums[,3:ncol(sums)] ~ Pond, sums, distance = "Bray")
adonis2(sums[,3:ncol(sums)] ~ as.numeric(Year), sums, distance = "Bray")

```
SIMPER ANALYSIS

```{r}
#Year-to-Centroid
##FTs
#colnames(FT.sums)[15] <- "N_P"
pond_avg_FTs <- FT.sums %>%
  group_by(Pond)%>%
  summarise(Osf = mean(Osf), Osh = mean(Osh), Ops = mean(Ops), Pro = mean(Pro),
            ES = mean(Osh), EP = mean(Osh), Bsh = mean(Bsh), Bsf = mean(Bsf),
            Pfp = mean(Pfp), Cpt = mean(Cpt), NH4 = mean(NH4), PO4 = mean(PO4),
            N_P = mean(N_P))

pond_avg_FTs <- cbind(pond_avg_FTs[,1], rep("Avg", 6), pond_avg_FTs[,2:14])
colnames(pond_avg_FTs)[2] <- "Year"
Yr_cent_FTs <- rbind(pond_avg_FTs, FT.sums)


uni.yr_cents <- c("2010-avg", "2011-avg", "2012-avg","2013-avg", "2014-avg", "2015-avg")

avg_yr_cent_FTs <-data.frame(FT = colnames(FT.sums[,3:ncol(FT.sums)]))
for(i in 1:length(ponds)){
  tmp <- subset(Yr_cent_FTs, Pond == ponds[i])
  tmp <- rbind(tmp, tmp)
  simp_yr_cent_FT <- simper(tmp[,3:ncol(tmp)], tmp$Year)
  for(j in 1:length(uni.yr_cents)){
    tmp2 <- as.data.frame(simp_yr_cent_FT[[j]]$average)
    colnames(tmp2)[1] <- paste(ponds[i], uni.yr_cents[j], sep = "_")
    avg_yr_cent_FTs <- cbind(avg_yr_cent_FTs, tmp2)
  }
}
avg_yr_cent_FTs[,2:ncol(avg_yr_cent_FTs)] <- round(avg_yr_cent_FTs[,2:ncol(avg_yr_cent_FTs)],3)


##COMM 
pond_avg_coms <- sums %>%
  pivot_longer(3:ncol(sums),
               names_to = "species",
               values_to = "count") %>%
  group_by(Pond, species) %>%
  summarise(mean_count = mean(count)) %>%
  pivot_wider(names_from = "species",
              values_from = "mean_count")

pond_avg_coms <- cbind(pond_avg_coms[,1], rep("Avg", 6), pond_avg_coms[,2:28])
colnames(pond_avg_coms)[2] <- "Year"
sums$Year <- as.character(sums$Year)
Yr_cent_coms <- rbind(pond_avg_coms, sums)



avg_yr_cent_coms<- data.frame(spc = colnames(sums[,3:ncol(sums)]))
for(i in 1:length(ponds)){
  tmp <- subset(Yr_cent_coms, Pond == ponds[i])
  tmp <- rbind(tmp, tmp)
  simp_yr_cent_com <- simper(tmp[,3:ncol(tmp)], tmp$Year)
  
  for(j in 1:length(uni.yr_cents)){
    tmp2 <- as.data.frame(simp_yr_cent_com[[j]]$average)
    colnames(tmp2)[1] <- paste(ponds[i], uni.yr_cents[j], sep = "_")
    avg_yr_cent_coms <- cbind(avg_yr_cent_coms, tmp2)
  }
}

avg_yr_cent_coms[,2:ncol(avg_yr_cent_coms)] <- round(avg_yr_cent_coms[,2:ncol(avg_yr_cent_coms)],3)

#Year-to-Year

uni.yr_yr <- c("2010-2011", "2011-2012", "2012-2013","2013-2014", "2014-2015")
uni.y_yr_index <- c(1, 6, 10, 13, 15)
##FTs
avg_yr_yr_FTs <- data.frame(FT = colnames(FT.sums[,3:ncol(FT.sums)]))
for(i in 1:length(ponds)){
  tmp <- subset(FT.sums, Pond == ponds[i])
  tmp <- rbind(tmp, tmp)
  simp_yr_yr_FT <- simper(tmp[,3:ncol(tmp)], tmp$Year)
  for(j in 1:length(uni.yr_yr)){
    tmp2 <- as.data.frame(simp_yr_yr_FT[[uni.y_yr_index[j]]]$average)
    colnames(tmp2)[1] <- paste(ponds[i], uni.yr_yr[j], sep = "_")
    avg_yr_yr_FTs <- cbind(avg_yr_yr_FTs, tmp2)
  }
}
avg_yr_yr_FTs[,2:ncol(avg_yr_yr_FTs)] <- round(avg_yr_yr_FTs[,2:ncol(avg_yr_yr_FTs)],3)


##COMM 
avg_yr_yr_coms <- data.frame(spc= colnames(sums[,3:ncol(sums)]))
for(i in 1:length(ponds)){
  tmp <- subset(sums, Pond == ponds[i])
  tmp <- rbind(tmp, tmp)
  simp_yr_yr_com <- simper(tmp[,3:ncol(tmp)], tmp$Year)
  for(j in 1:length(uni.yr_yr)){
    tmp2 <- as.data.frame(simp_yr_yr_com[[uni.y_yr_index[j]]]$average)
    colnames(tmp2)[1] <- paste(ponds[i], uni.yr_yr[j], sep = "_")
    avg_yr_yr_coms <- cbind(avg_yr_yr_coms, tmp2)
  }
}
avg_yr_yr_coms[,2:ncol(avg_yr_yr_coms)] <- round(avg_yr_yr_coms[,2:ncol(avg_yr_yr_coms)],3)

```
#Plot Simper 
```{r}
spcs <- data.frame(Common_name = colnames(sums[,3:ncol(sums)]))
#spcs$Common_name <- gsub(".", " ", spcs$Common_name, fixed = T)
co.FT <- as.data.frame(c("#a3c4af","#80ae90","#5c9a71","#89c9d1","#449ca4","#0e7e8a","#125361","#0d373e","#ebcc8f","#e5bd75","#ddad58","#d59d3b","#ad7b26"))
co.FT$FT <-  c("N_P", "NH4", "PO4","Bsf", "Bsh", "Cpt", "Pfp","EP", "ES","Ops","Osh", "Osf","Pro")
colnames(co.FT)[1] <- "col"

 
col <- pnw_palette("Sunset",100)
col.detrit <- col[c(1,7,13)]
col.plankt <- col[c(20,26,32,38,44,46)]
col.meio<- col[c(55,57,61,63,65,67,69,71,73)]
col.macro<- col[c(84,86,88,90,92,94,96,98,100)]


TL <- merge(TL, spcs,  by = "Common_name")

 TL <-TL %>%
   arrange(factor(Feeding_habit, levels = c("detritovore", "plankton feeding", "hunting meiofauna", "hunting macrofauna")))
  TL$col <- c(col.detrit, col.plankt, col.meio, col.macro)
  TL$piecol[TL$Feeding_habit== "detritovore"] <- "#4E4B6F"  
  TL$piecol[TL$Feeding_habit == "plankton feeding"] <- "#835B75" 
  TL$piecol[TL$Feeding_habit == "hunting meiofauna"] <- "#DB9770"
  TL$piecol[TL$Feeding_habit == "hunting macrofauna"] <- "#F0C186"  
  
  
  
#YR to Avg
#FTS
  avg_yr_cent_FTs$FT <- factor(avg_yr_cent_FTs$FT, levels = co.FT$FT)

avg_yr_cent_FTs_lon <- avg_yr_cent_FTs %>%
  pivot_longer(cols = 2:ncol(avg_yr_cent_FTs),
               names_to = "pond_yr",
              values_to = "avg")
avg_yr_cent_FTs_lon$Pond <- substr(avg_yr_cent_FTs_lon$pond_yr, 1,2)
avg_yr_cent_FTs_lon$year <- substr(avg_yr_cent_FTs_lon$pond_yr, 4,11)

ggplot(avg_yr_cent_FTs_lon, aes(x = year)) +
  geom_bar(aes(fill = FT, y = avg), position = "stack", stat = "identity") +
  theme_classic() +
  facet_wrap(.~Pond)+
scale_fill_manual(values = as.character(co.FT$col)) +
      scale_x_discrete(guide = guide_axis(angle = 90)) +
  theme(axis.text = element_text(size = 18),
        axis.title = element_text(size = 20)) +
  ylab("Contribution to dissimilarity") +
  xlab("Year to Average")

       
#species
avg_yr_cent_coms$spc <- factor(avg_yr_cent_coms$spc, levels = TL$Common_name)

avg_yr_cent_coms_lon <- avg_yr_cent_coms %>%
  pivot_longer(cols = 2:ncol(avg_yr_cent_coms),
               names_to = "pond_yr",
              values_to = "avg")
avg_yr_cent_coms_lon$Pond <- substr(avg_yr_cent_coms_lon$pond_yr, 1,2)
avg_yr_cent_coms_lon$year <- substr(avg_yr_cent_coms_lon$pond_yr, 4,11)

avg_yr_cent_coms_lon$spc <- factor(avg_yr_cent_coms_lon$spc, levels = TL$Common_name)

ggplot(avg_yr_cent_coms_lon, aes(x = year)) +
  geom_bar(aes(fill = spc, y = avg), position = "stack", stat = "identity") +
  theme_classic() +
  facet_wrap(.~Pond)+
scale_fill_manual(values = as.character(TL$col)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
  #theme(axis.text = element_text(size = 18),
   #     axis.title = element_text(size = 20)) +
  ylab("Contribution to dissimilarity") +
  xlab("Year to centroid")
  theme(legend.position="none")


#YR to YR
#FTS
avg_yr_yr_FTs$FT <- factor(avg_yr_yr_FTs$FT, levels = co.FT$FT)

avg_yr_yr_FTs_lon <- avg_yr_yr_FTs %>%
  pivot_longer(cols = 2:ncol(avg_yr_yr_FTs),
               names_to = "pond_yr",
              values_to = "avg")
avg_yr_yr_FTs_lon$Pond <- substr(avg_yr_yr_FTs_lon$pond_yr, 1,2)
avg_yr_yr_FTs_lon$year <- substr(avg_yr_yr_FTs_lon$pond_yr, 4,12)

ggplot(avg_yr_yr_FTs_lon, aes(x = year)) +
  geom_bar(aes(fill = FT, y = avg), position = "stack", stat = "identity") +
  theme_classic() +
  facet_wrap(.~Pond)+
scale_fill_manual(values = as.character(co.FT$col)) +
 # theme(axis.text = element_text(size = 18),
 #       axis.title = element_text(size = 20)) +
      scale_x_discrete(guide = guide_axis(angle = 90)) +
  ylab("Contribution to dissimilarity") +
  xlab("Year to Average")


#species
avg_yr_yr_coms_lon <- avg_yr_yr_coms %>%
  pivot_longer(cols = 2:ncol(avg_yr_yr_coms),
               names_to = "pond_yr",
              values_to = "avg")
avg_yr_yr_coms_lon$Pond <- substr(avg_yr_yr_coms_lon$pond_yr, 1,2)
avg_yr_yr_coms_lon$year <- substr(avg_yr_yr_coms_lon$pond_yr, 4,12)

ggplot(avg_yr_yr_coms_lon, aes(x = year)) +
  geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
  theme_classic() +
  facet_wrap(.~Pond)+
scale_fill_manual(values = as.character(TL$col)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
  theme(axis.text = element_text(size = 18),
        axis.title = element_text(size = 20)) +
  ylab("Contribution to dissimilarity") +
  xlab("Year to Year")
```

#summed contribution of each FT or hunting mode type 
```{r}

co.FT$type <- c(rep("Nutrient recycling",3), rep("Locomotion",5), rep("Energy Acquisition", 5))
co.FT$col_type <- c(rep("#80ae90",3), rep("#0e7e8a",5), rep("#ddad58",5))
#Yr to cent
#FTs
FT_pond_contrib_yr_cent <- avg_yr_cent_FTs_lon %>%
  left_join(co.FT[,2:3], by = "FT")%>%
  group_by(Pond, type) %>%
  summarize(sum_contrib = sum(avg))

#Species
com_pond_contrib_yr_cent <- avg_yr_cent_coms_lon %>%
  left_join(TL[,c(1,5)],join_by("spc" == "Common_name"))%>%
  group_by(Pond, Feeding_habit) %>%
  summarize(sum_contrib = sum(avg))

#Yr to yr
#FTs
FT_pond_contrib_yr_yr <- avg_yr_yr_FTs_lon %>%
  left_join(co.FT[,2:3], by = "FT")%>%
  group_by(Pond, type) %>%
  summarize(sum_contrib = sum(avg))

#Species
com_pond_contrib_yr_yr <- avg_yr_yr_coms_lon %>%
  left_join(TL[,c(1,5)],join_by("spc" == "Common_name"))%>%
  group_by(Pond, Feeding_habit) %>%
  summarize(sum_contrib = sum(avg))


#PIE plot theme
blank_theme <- theme_minimal()+
  theme(
  axis.title.x = element_blank(),
  axis.title.y = element_blank(),
  panel.border = element_blank(),
  panel.grid=element_blank(),
  axis.ticks = element_blank(),
  plot.title=element_text(size=14, face="bold")
  )
```

Making final simper plots
COMs Year to centroid
```{r}
com_pond_contrib_yr_cent <- com_pond_contrib_yr_cent %>%
  arrange(factor(Feeding_habit, levels = c("detritovore", "plankton feeding", "hunting meiofauna", "hunting macrofauna")))

avg_yr_cent_coms_lon <- avg_yr_cent_coms_lon %>%
  arrange(factor(spc, levels = TL$Common_name))
   
tmp <- subset(com_pond_contrib_yr_cent, Pond == ponds[1])
(b1 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_coms_lon, Pond == ponds[1])

(f1 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), 
             position = "stack", stat = "identity") +
    ylim(0,0.32)+
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
   scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("")+
    annotate("text", x = 0.75, y = 0.32, label = ponds[1], size = 5) + 
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b1),  xmin = 5.4, xmax = 6.8, 
                    ymin = 0.22, ymax = 0.38) + labs(fill = 'spc'))

tmp <- subset(com_pond_contrib_yr_cent, Pond == ponds[2])
(b2 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))

tmp2 <- subset(avg_yr_cent_coms_lon, Pond == ponds[2])

(f2 <-   ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylim(0,0.75)+
    ylab("") +
    xlab("")+
    annotate("text", x = 0.75, y = 0.75, label = ponds[2], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 5.3, xmax = 6.7,
                    ymin = 0.62, ymax = 0.85) + labs(fill='spc'))

tmp <- subset(com_pond_contrib_yr_cent, Pond == ponds[3])
(b3 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_coms_lon, Pond == ponds[3])

(f3 <-ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("")+
    annotate("text", x = 0.75, y = 0.54, label = ponds[3], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b3),  xmin = 5.3, xmax = 6.8, 
                    ymin = 0.43, ymax = 0.62) + labs(fill='spc'))

tmp <- subset(com_pond_contrib_yr_cent, Pond == ponds[4])
(b4 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_coms_lon, Pond == ponds[4])

(f4 <-ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("Year to Centroid")+
    annotate("text", x = 0.85, y = 0.62, label = ponds[4], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b4), xmin = 5.2, xmax = 6.8,
                      ymin = 0.48, ymax = 0.69) + labs(fill='spc'))


tmp <- subset(com_pond_contrib_yr_cent, Pond == ponds[5])
(b5 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_coms_lon, Pond == ponds[5])

(f5 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("Year to Centroid")+
    annotate("text", x = 0.8, y = 0.82, label = ponds[5], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b5),  xmin = 5, xmax = 7,
                      ymin = 0.66, ymax = 0.92) + labs(fill='spc'))

tmp <- subset(com_pond_contrib_yr_cent, Pond == ponds[6])
(b6 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75" ))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_coms_lon, Pond == ponds[6])

(f6 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("Year to Centroid")+
    annotate("text", x = 0.81, y = 1, label = ponds[6], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b6), xmin = 5, xmax = 7,
                      ymin = 0.81, ymax = 1.15) + labs(fill='spc'))


(f1 | f2 | f3)/
(f4 | f5 | f6) + plot_layout(guides = "collect")  & theme( axis.text = element_text(size =  11), axis.title = element_text(size = 14))
#ggsave("SIMPER analysis/Plots/simper_com_yr_cent.png")
```

COMs Year to year
```{r}

com_pond_contrib_yr_yr <- com_pond_contrib_yr_yr %>%
  arrange(factor(Feeding_habit, levels = c("detritovore", "plankton feeding", "hunting meiofauna", "hunting macrofauna")))

avg_yr_yr_coms_lon <- avg_yr_yr_coms_lon %>%
  arrange(factor(spc, levels = TL$Common_name))

tmp <- subset(com_pond_contrib_yr_yr, Pond == ponds[1])
(b1 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_coms_lon, Pond == ponds[1])

(f1 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("")+
    annotate("text", x = 0.81, y = 0.551, label = ponds[1], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b1),  xmin = 4, xmax = 6,
                    ymin = 0.45, ymax = 0.65) + labs(fill='spc'))

tmp <- subset(com_pond_contrib_yr_yr, Pond == ponds[2])
(b2 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))

tmp2 <- subset(avg_yr_yr_coms_lon, Pond == ponds[2])

(f2 <-   ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("")+
    annotate("text", x = 0.81, y = 1.1, label = ponds[2], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 3.5, xmax = 6.5,
                    ymin = 0.82, ymax = 1.22) + labs(fill='spc'))

tmp <- subset(com_pond_contrib_yr_yr, Pond == ponds[3])
(b3 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_coms_lon, Pond == ponds[3])

(f3 <-ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("")+
    annotate("text", x = 0.81, y = 0.8, label = ponds[3], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b3),  xmin = 3.5, xmax = 6.5,
                    ymin = 0.62, ymax = 0.92) +labs(fill='spc'))

tmp <- subset(com_pond_contrib_yr_yr, Pond == ponds[4])
(b4 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_coms_lon, Pond == ponds[4])

(f4 <-ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("Year to Year")+
    theme(legend.position = "none")+ 
    annotate("text", x = 0.81, y = 0.8, label = ponds[4], size = 5) +
    annotation_custom(ggplotGrob(b4), xmin = 3.5, xmax = 6.5,
                    ymin = 0.62, ymax = 0.92)  + labs(fill='spc'))


tmp <- subset(com_pond_contrib_yr_yr, Pond == ponds[5])
(b5 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_coms_lon, Pond == ponds[5])

(f5 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("Year to Year")+
    annotate("text", x = 0.81, y = 1.2, label = ponds[5], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b5), xmin = 3.5, xmax = 6.5,
                    ymin = 0.9, ymax = 1.4) + labs(fill='spc'))

tmp <- subset(com_pond_contrib_yr_yr, Pond == ponds[6])
(b6 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = Feeding_habit)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#4E4B6F", "#F0C186","#DB9770", "#835B75" ))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_coms_lon, Pond == ponds[6])

(f6 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("Year to Year")+
    annotate("text", x = 0.81, y = 1.2, label = ponds[6], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b6), xmin = 3.5, xmax = 6.5,
                    ymin = 0.9, ymax = 1.4) + labs(fill='spc'))



(f1 | f2 | f3)/
(f4 | f5 | f6) + plot_layout(guides = "collect")   & theme( axis.text = element_text(size =  11), axis.title = element_text(size = 14))

#ggsave("SIMPER analysis/Plots/simper_com_yr_yr.png")

#species legend
tmp2 <- subset(avg_yr_yr_coms_lon, Pond == ponds[6])

ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(spc, levels = TL$Common_name), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(TL$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) 
#ggsave("SIMPER analysis/Plots/community_legend.png")
```

FTs Year to centroid
```{r}
tmp <- subset(FT_pond_contrib_yr_cent, Pond == ponds[1])
(b1 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_FTs_lon, Pond == ponds[1])

(f1 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack",
             stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("")+
    annotate("text", x = 0.81, y = 0.12, label = ponds[1], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b1), xmin = 4.4, xmax = 7.7,
                      ymin = 0.09, ymax = 0.13) + labs(fill='FTs'))

tmp <- subset(FT_pond_contrib_yr_cent, Pond == ponds[2])
(b2 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_FTs_lon, Pond == ponds[2])

(f2 <-  ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("")+
    annotate("text", x = 0.81, y = 0.25, label = ponds[2], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.5, xmax = 7.7,
                      ymin = 0.19, ymax = 0.28) + labs(fill='FTs'))

tmp <- subset(FT_pond_contrib_yr_cent, Pond == ponds[3])
(b3 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_FTs_lon, Pond == ponds[3])

(f3 <-ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
   scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("")+
    annotate("text", x = 0.81, y = 0.14, label = ponds[3], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.6, xmax = 7.5,
                      ymin = 0.1, ymax = 0.155) + labs(fill='FTs'))

tmp <- subset(FT_pond_contrib_yr_cent, Pond == ponds[4])
(b4 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_FTs_lon, Pond == ponds[4])

(f4 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
   scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("Year to Centroid")+
    annotate("text", x = 0.88, y = 0.2, label = ponds[4], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.55, xmax = 7.6,
                      ymin = 0.14, ymax = 0.215) + labs(fill='FTs'))


tmp <- subset(FT_pond_contrib_yr_cent, Pond == ponds[5])
(b5 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_FTs_lon, Pond == ponds[5])

(f5 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
   scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("Year to Centroid")+
    annotate("text", x = 0.84, y = 0.25, label = ponds[5], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.5, xmax = 7.7,
                      ymin = 0.19, ymax = 0.28) + labs(fill='FTs'))

tmp <- subset(FT_pond_contrib_yr_cent, Pond == ponds[6])
(b6 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_cent_FTs_lon, Pond == ponds[6])

(f6 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("Year to Centroid")+
    annotate("text", x = 0.84, y = 0.25, label = ponds[6], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.5, xmax = 7.7,
                      ymin = 0.19, ymax = 0.28) + labs(fill='FTs'))


(f1 | f2 | f3)/
(f4 | f5 | f6) + plot_layout(guides = "collect")  & theme(legend.position = "right", axis.text = element_text(size = 11), axis.title = element_text(size = 14))
#ggsave("SIMPER analysis/Plots/simper_FT_yr_cent.png")
```

FTs Year to year
```{r}
tmp <- subset(FT_pond_contrib_yr_yr, Pond == ponds[1])
(b1 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_FTs_lon, Pond == ponds[1])

(f1 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("")+
    theme(legend.position = "none")+ 
    annotate("text", x = 0.84, y = 0.15, label = ponds[1], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.2, xmax = 5.7,
                      ymin = 0.09, ymax = 0.2) + labs(fill='FTs'))

tmp <- subset(FT_pond_contrib_yr_yr, Pond == ponds[2])
(b2 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_FTs_lon, Pond == ponds[2])

(f2 <-  ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("")+
    annotate("text", x = 0.84, y = 0.35, label = ponds[2], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2), xmin = 4.2, xmax = 5.7,
                      ymin = 0.25, ymax = 0.4) + labs(fill='FTs'))

tmp <- subset(FT_pond_contrib_yr_yr, Pond == ponds[3])
(b3 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_FTs_lon, Pond == ponds[3])

(f3 <-ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("")+
    annotate("text", x = 0.84, y = 0.18, label = ponds[3], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.2, xmax = 5.7,
                      ymin = 0.10, ymax = 0.23)+ labs(fill='FTs'))


tmp <- subset(FT_pond_contrib_yr_yr, Pond == ponds[4])
(b4 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_FTs_lon, Pond == ponds[4])

(f4 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("Contribution to dissimilarity") +
    xlab("Year to Year")+
    annotate("text", x = 0.84, y = 0.3, label = ponds[4], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),   xmin = 4.2, xmax = 5.7,
                      ymin = 0.22, ymax = 0.33) + labs(fill='FTs'))


tmp <- subset(FT_pond_contrib_yr_yr, Pond == ponds[5])
(b5 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_FTs_lon, Pond == ponds[5])

(f5 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    ylab("") +
    xlab("Year to Year")+
    annotate("text", x = 0.84, y = 0.35, label = ponds[5], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.2, xmax = 5.7,
                      ymin = 0.27, ymax = 0.42) + labs(fill='FTs'))

tmp <- subset(FT_pond_contrib_yr_yr, Pond == ponds[6])
(b6 <- ggplot(tmp, aes(x = "", y = sum_contrib, fill = type)) +
  geom_bar(width = 1, stat = "identity") + coord_polar("y", start = 0) + blank_theme +
   scale_fill_manual(values = as.character(c("#ddad58","#0e7e8a", "#80ae90"))) +
 theme(axis.text.x = element_blank())  +
    theme(legend.position = "none"))
tmp2 <- subset(avg_yr_yr_FTs_lon, Pond == ponds[6])

(f6 <- ggplot(tmp2, aes(x = year)) +
    geom_bar(aes(fill = factor(FT, levels = co.FT$FT), y = avg), position = "stack", stat = "identity") +
    theme_classic() +
    scale_fill_manual(values = as.character(co.FT$col)) +
    theme(axis.text = element_text(size = 16),
          axis.title = element_text(size = 18)) +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
   ylab("") +
    xlab("Year to Year")+
    annotate("text", x = 0.84, y = 0.3, label = ponds[6], size = 5) +
    theme(legend.position = "none")+ 
    annotation_custom(ggplotGrob(b2),  xmin = 4.2, xmax = 5.7,
                      ymin = 0.22, ymax = 0.37) + labs(fill='FTs'))


(f1 | f2 | f3)/
(f4 | f5 | f6) + plot_layout(guides = "collect")  & theme(legend.position = "right", axis.text = element_text(size = 11), axis.title = element_text(size = 14))
#ggsave("SIMPER analysis/Plots/simper_FT_yr_yr.png")
```


```{r ENV FIT BIPLOT}
TL <-TL %>%
    arrange(factor(Feeding_habit, levels = c("detritovore", "plankton feeding", "hunting meiofauna", "hunting macrofauna")))
   TL$col <- c(col.detrit, col.plankt, col.meio, col.macro)

fish.sums <- aggregate(FREQ ~ Common_name + Pond + Year, data = com.dat.fish, sum) 
TL.sums <- merge(TL[,c(1,4:5)], fish.sums, by = "Common_name")

avg.by.year <- aggregate(FREQ ~ Pond + Year + Feeding_habit, FUN = mean, data = TL.sums)
 sums.3 <- reshape(avg.by.year, timevar = c("Feeding_habit"), ids = c(FREQ), idvar = c("Pond","Year"), direction = "wide")
 sums.2 <- sums.3[,3:ncol(sums.3)]
 names(sums.2) <- substring(names(sums.2), 6)
sums.2[is.na(sums.2)] <- 0
 sums.TL <- cbind(sums.3[,1:2], sums.2)
 sums.TL <- sums.TL[order(sums.TL$Year),]
 sums.TL <- sums.TL[order(sums.TL$Pond),]

FT.mds <- metaMDS(FT.sums[,3:ncol(FT.sums)], k = 2, trymax = 10000, maxit = 10000)


col.FT <- as.data.frame(unique(FT.mds$species))
col.FT$FT <- rownames(col.FT)
col.FT[13,3] <- "N_P"
col.FT[11,3] <- "NH4"
col.FT[12,3] <- "PO4"

col.FT <- merge(col.FT, co.FT, by = "FT")

coords <- as.data.frame(FT.mds$points)
coords <- cbind(FT.sums[,1:2], coords)
coords <- merge(coords, pal, by = "Pond")

t_col <- function(color, percent = 100, name = NULL) {
  #	  color = color name
  #	percent = % transparency
  #	   name = an optional name for the color
  ## Get RGB values for named color
  rgb.val <- col2rgb(color)
  ## Make new color using input color as base and alpha set by transparency
  t.col <- rgb(rgb.val[1], rgb.val[2], rgb.val[3],
               max = 255,
               alpha = (100-percent)*255/100,
               names = name)
  ## Save the color
  invisible(t.col)
  
}


#ENV Fit correlating species with FTs
set.seed(20)
(com.fit <- envfit(FT.mds, sums[3:ncol(sums)], permutations = 9999))
com.fit.df <- as.data.frame(com.fit$vectors[c(1,2,4)])
com.fit.df <- cbind(rownames(com.fit.df), com.fit.df)
colnames(com.fit.df)[1] <- "Common_name"
com.fit.df <- merge(com.fit.df, TL[,c(1,5,15)], by = "Common_name" )
com.fit.df$t_col <- ifelse(com.fit.df$pvals < 0.07, 0, 
                                  ifelse(com.fit.df$pvals < 0.1, 20, 
                                        ifelse(com.fit.df$pvals < 0.4, 80, 100)))

#png("/Users/mallarie.yeager/Documents/R_projects/FT Stability/SIMPER analysis/Plots/biplot_species.png", width = 960, height = 960, units = "px")
#jpeg("04_ANCOVA_SIMPER_BIPOLT Analyses/Plots/biplot_species.jpeg", 
 #    width = 8.5, height = 8.5, units = "cm", res = 600, pointsize = 8)
plot(coords$MDS1, coords$MDS2,  type = "n", xlab = "nMDS1", xlim = c(-0.4,0.4), ylim = c(-0.3,0.3), ylab = "nMDS2", main = "")
abline(h = 0, lty = 2)
abline(v = 0, lty = 2)
text(col.FT$MDS1, col.FT$MDS2, labels = col.FT$FT,col = col.FT$col, cex = 1)
#text(-0.51,0.3,labels = "(a)")
#plot(com.fit, col = com.fit.df$col)
plot(com.fit, col= c(t_col(com.fit.df$col[1], com.fit.df$t_col[1]),
                     t_col(com.fit.df$col[2], com.fit.df$t_col[2]),
                     t_col(com.fit.df$col[3], com.fit.df$t_col[3]),
                     t_col(com.fit.df$col[4], com.fit.df$t_col[4]),
                     t_col(com.fit.df$col[5], com.fit.df$t_col[5]),
                     t_col(com.fit.df$col[6], com.fit.df$t_col[6]),
                     t_col(com.fit.df$col[7], com.fit.df$t_col[7]),
                     t_col(com.fit.df$col[8], com.fit.df$t_col[8]),
                     t_col(com.fit.df$col[9], com.fit.df$t_col[9]),
                     t_col(com.fit.df$col[10], com.fit.df$t_col[10]),
                     t_col(com.fit.df$col[11], com.fit.df$t_col[11]),
                     t_col(com.fit.df$col[12], com.fit.df$t_col[12]),
                     t_col(com.fit.df$col[13], com.fit.df$t_col[13]),
                     t_col(com.fit.df$col[14], com.fit.df$t_col[14]),
                     t_col(com.fit.df$col[15], com.fit.df$t_col[15]),
                     t_col(com.fit.df$col[16], com.fit.df$t_col[16]),
                     t_col(com.fit.df$col[17], com.fit.df$t_col[17]),
                     t_col(com.fit.df$col[18], com.fit.df$t_col[18]),
                     t_col(com.fit.df$col[19], com.fit.df$t_col[19]),
                     t_col(com.fit.df$col[20], com.fit.df$t_col[20]),
                     t_col(com.fit.df$col[21], com.fit.df$t_col[21]),
                     t_col(com.fit.df$col[22], 100),
                     t_col(com.fit.df$col[23], com.fit.df$t_col[23]),
                     t_col(com.fit.df$col[24], com.fit.df$t_col[24]),
                     t_col(com.fit.df$col[25], com.fit.df$t_col[25]),
                     t_col(com.fit.df$col[26], com.fit.df$t_col[26]),
                     t_col(com.fit.df$col[27], com.fit.df$t_col[27])), 
                     cex = 0.9, lwd = 6)
text(-0.155, -0.23, labels = "Striped Killifish", col = com.fit.df$col[22], cex = 0.9)
legend("topleft", legend = c("Detritovore", "Planktivore", "Hunting Meiofauna", "Hunting Macrofauna"), col = c(as.character("#4E4B6F"), as.character("#835B75"), as.character("#DB9770"), as.character("#F0C186")), 
       lty = 1, lwd = 3, bty= "n", cex = 0.5)
#dev.off()


#ENV Fit correlating trophic groups with FTs
set.seed(22)
 (feed.fit <- envfit(FT.mds, sums.TL[,3:ncol(sums.TL)], permutations = 9999))
 feed.fit.df <- as.data.frame(feed.fit$vectors[c(1,2,4)])
feed.fit.df <- cbind(rownames(feed.fit.df), feed.fit.df)
colnames(feed.fit.df)[1] <- "Feeding_habit"
feed.fit.df$col <- c("#4E4B6F", "#F0C186", "#DB9770","#835B75" )
feed.fit.df$t_col <- ifelse(feed.fit.df$pvals < 0.07, 0, 
                                  ifelse(feed.fit.df$pvals < 0.1, 20, 
                                        ifelse(feed.fit.df$pvals < 0.4, 80, 100)))


#jpeg("04_ANCOVA_SIMPER_BIPOLT Analyses/Plots/biplot_feeding_habit.jpeg", width = 8.5, height = 8.5, units = "cm", res = 400)
plot(coords$MDS1, coords$MDS2,  type = "n", xlab = "nMDS1", xlim = c(-0.4,0.4), ylim = c(-0.3,0.3), ylab = "nMDS2", main = "")
abline(h = 0, lty = 2)
abline(v = 0, lty = 2)
text(col.FT$MDS1, col.FT$MDS2, labels = col.FT$FT,col = col.FT$col)
plot(feed.fit, col= c(t_col(feed.fit.df$col[1], feed.fit.df$t_col[1]),
                     t_col(feed.fit.df$col[2], feed.fit.df$t_col[2]),
                     t_col(feed.fit.df$col[3], feed.fit.df$t_col[3]),
                     t_col(feed.fit.df$col[4], feed.fit.df$t_col[4])), lwd = 6)

legend("topright", legend = c("Detritovore", "Planktivore", "Hunting Meiofauna", "Hunting Macrofauna"), col = c(as.character("#4E4B6F"), as.character("#835B75"), as.character("#DB9770"), as.character("#F0C186")), 
       lty = 1, lwd = 3, bty= "n", cex = 0.5)
#dev.off()
```


Figure 4 for manuscript
```{r}
jpeg("04_ANCOVA_SIMPER_BIPOLT Analyses/Plots/biplot_species.jpeg", 
     width = 8.5, height = 8.5, units = "cm", res = 600, pointsize = 8)
plot(coords$MDS1, coords$MDS2,  type = "n", xlab = "nMDS1", xlim = c(-0.4,0.4), ylim = c(-0.3,0.3), ylab = "nMDS2", main = "")
abline(h = 0, lty = 2)
abline(v = 0, lty = 2)
text(col.FT$MDS1, col.FT$MDS2, labels = col.FT$FT,col = col.FT$col, cex = 1)
#text(-0.51,0.3,labels = "(a)")
#plot(com.fit, col = com.fit.df$col)
plot(com.fit, col= c(t_col(com.fit.df$col[1], com.fit.df$t_col[1]),
                     t_col(com.fit.df$col[2], com.fit.df$t_col[2]),
                     t_col(com.fit.df$col[3], com.fit.df$t_col[3]),
                     t_col(com.fit.df$col[4], com.fit.df$t_col[4]),
                     t_col(com.fit.df$col[5], com.fit.df$t_col[5]),
                     t_col(com.fit.df$col[6], com.fit.df$t_col[6]),
                     t_col(com.fit.df$col[7], com.fit.df$t_col[7]),
                     t_col(com.fit.df$col[8], com.fit.df$t_col[8]),
                     t_col(com.fit.df$col[9], com.fit.df$t_col[9]),
                     t_col(com.fit.df$col[10], com.fit.df$t_col[10]),
                     t_col(com.fit.df$col[11], com.fit.df$t_col[11]),
                     t_col(com.fit.df$col[12], com.fit.df$t_col[12]),
                     t_col(com.fit.df$col[13], com.fit.df$t_col[13]),
                     t_col(com.fit.df$col[14], com.fit.df$t_col[14]),
                     t_col(com.fit.df$col[15], com.fit.df$t_col[15]),
                     t_col(com.fit.df$col[16], com.fit.df$t_col[16]),
                     t_col(com.fit.df$col[17], com.fit.df$t_col[17]),
                     t_col(com.fit.df$col[18], com.fit.df$t_col[18]),
                     t_col(com.fit.df$col[19], com.fit.df$t_col[19]),
                     t_col(com.fit.df$col[20], com.fit.df$t_col[20]),
                     t_col(com.fit.df$col[21], com.fit.df$t_col[21]),
                     t_col(com.fit.df$col[22], 100),
                     t_col(com.fit.df$col[23], com.fit.df$t_col[23]),
                     t_col(com.fit.df$col[24], com.fit.df$t_col[24]),
                     t_col(com.fit.df$col[25], com.fit.df$t_col[25]),
                     t_col(com.fit.df$col[26], com.fit.df$t_col[26]),
                     t_col(com.fit.df$col[27], com.fit.df$t_col[27])), 
                     cex = 0.9, lwd = 6)
text(-0.155, -0.23, labels = "Striped Killifish", col = com.fit.df$col[22], cex = 0.9)
legend("topleft", legend = c("Detritovore", "Planktivore", "Hunting Meiofauna", "Hunting Macrofauna"), col = c(as.character("#4E4B6F"), as.character("#835B75"), as.character("#DB9770"), as.character("#F0C186")), 
       lty = 1, lwd = 3, bty= "n", cex = 0.5)
dev.off()

```
